### 2020-12-08 #2895 @whedon
[PRE REVIEW]: Moead-framework : a modular MOEA/D python framework
**Submitting author:** @geoffreyp (<a href="http://orcid.org/0000-0002-5049-5991">Geoffrey Pruvost</a>)
**Repository:** <a href="https://github.com/moead-framework/framework" target ="_blank">https://github.com/moead-framework/framework</a>
**Version:** 0.5.7.1
