### 2020-12-17 #2895 @kthyng
Hi @geoffreyp, your paper is missing a section titled "Statement of Need." It also should start with a high level scientific summary that a scientific person not in your area would understand.
