### 2020-12-18 #2895 @geoffreyp
Hello @kthyng, thank you for your feedback. The paper is updated with the "Statement of Need" section.
