### 2021-01-20 #2974 @whedon
[REVIEW]: Moead-framework : a modular MOEA/D python framework
**Submitting author:** @geoffreyp (<a href="http://orcid.org/0000-0002-5049-5991">Geoffrey Pruvost</a>)
**Repository:** https://github.com/moead-framework/framework
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
