+++ moead_framework/core/sps_strategy/__init__.py
@@ -1,0 +1,3 @@
+from moead_framework.core.sps_strategy.sps_all import SpsAllSubproblems
+from moead_framework.core.sps_strategy.sps_dra import SpsDra
+from moead_framework.core.sps_strategy.sps_random_and_boundaries import SpsRandomAndBoundaries
