+++ moead_framework/algorithm/numerical/__init__.py
@@ -1,0 +1,1 @@
+from moead_framework.algorithm.numerical.moead import Moead
