+++ example_rmnk_moead.py
@@ -1,20 +1,20 @@
-from moead_framework.aggregation.tchebycheff import Tchebycheff
-from moead_framework.algorithm.combinatorial.moead import Moead
-from moead_framework.problem.combinatorial.rmnk import Rmnk
+from moead_framework.aggregation import Tchebycheff
+from moead_framework.algorithm.combinatorial import Moead
+from moead_framework.problem.combinatorial import Rmnk
 from moead_framework.tool.result import save_population
 
 
 ###############################
-#      Init the problem       #
+#   Initialize the problem    #
 ###############################
 # Others instances are available here : https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances
 instance_file = "moead_framework/test/data/RMNK/Instances/rmnk_0_2_100_1_0.dat"
 rmnk = Rmnk(instance_file=instance_file)
 
 
-###############################
-#      Init the algorithm     #
-###############################
+#####################################
+#      Initialize the algorithm     #
+#####################################
 number_of_objective = rmnk.function_numbers
 number_of_weight = 10
 number_of_weight_neighborhood = 2
