+++ moead_framework/core/selector/__init__.py
@@ -1,0 +1,2 @@
+from moead_framework.core.selector.closest_neighbors_selector import ClosestNeighborsSelector
+from moead_framework.core.selector.delta_selector import DeltaSelector
