### 2021-02-06 moead-framework/framework#10 @chkoar
[joss-reviews#2974] Python and dependencies version compatibility are missing
Related: https://github.com/openjournals/joss-reviews/issues/2974

Docs should state the minimum (and the expected maximum) required versions so the package would be able to work.
