### 2021-02-06 moead-framework/framework#11 @chkoar
[joss-reviews#2974] CI should run against the required versions
Related: https://github.com/openjournals/joss-reviews/issues/2974, #10
