### 2021-02-06 moead-framework/framework#12 @chkoar
[joss-reviews#2974] Improve the test coverage
Related: https://github.com/openjournals/joss-reviews/issues/2974

Some modules have pretty low coverage. Since you are working in a research context consider to increase the test coverage so the software could be trusted. It would be nice to provide a badge in the read me regarding the code coverage.

Below I list some files:

```
Name                                                                                  Stmts   Miss  Cover   Missing
-------------------------------------------------------------------------------------------------------------------
moead_framework\tool\result.py                                                          249    151    39%   5-24, 28-38, 42-52, 56-69, 73-75, 80-89, 93-100, 104-109, 146, 175, 179, 195-243, 265, 268, 288, 299-310, 314, 318-324, 328-333, 348-355, 362-366
moead_framework\core\genetic_operator\numerical\moead_de_operators.py                    14      7    50%   12-22
moead_framework\tool\mop.py                                                              68     24    65%   25-27, 40, 78-88, 110-124
moead_framework\solution\one_dimension_solution.py                                       15      4    73%   13, 16, 19, 22
moead_framework\core\sps_strategy\sps_dra.py                                             34      9    74%   37-47
moead_framework\aggregation\functions.py                                                  8      2    75%   19, 29
moead_framework\problem\problem.py                                                       13      3    77%   12, 16, 20
moead_framework\core\genetic_operator\abstract_operator.py                                9      2    78%   11, 15
```
