### 2021-02-06 moead-framework/framework#13 @chkoar
[joss-reviews#2974] Add community guidelines
Related: openjournals/joss-reviews#2974

See [here](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines).
