### 2021-02-06 moead-framework/framework#14 @chkoar
[joss-reviews#2974] Consider adding a user guide
Related: openjournals/joss-reviews#2974

Currently the documentation seems like a single page. Please consider adding a user guide. See [here](https://deap.readthedocs.io/en/master/) for instance.
