### 2021-02-06 moead-framework/framework#15 @chkoar
[joss-reviews#2974] Improve API docs
Related: openjournals/joss-reviews#2974

All functions/methods should be documented including example inputs and outputs and their expected types. [This](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html) may help you.
