### 2021-02-06 moead-framework/framework#16 @chkoar
Question regarding the coupling between problems and files
`Related: openjournals/joss-reviews#2974`

Why the problem classes are coupled with files and not with other data structures?
