### 2021-02-06 moead-framework/framework#17 @chkoar
[joss-reviews#2974] Class design regarding Moead and Problems
Related: openjournals/joss-reviews#2974

`Moead` and `Problem` classes require files, but I cannot see their definition somewhere documented.

A better option could be to work directly with data structures. (arrays, lists, dicts)
Helper functions could be created in order to avoid the clutter of loading these.
