### 2021-02-06 moead-framework/framework#18 @chkoar
[joss-reviews#2974] Code duplication
Related: openjournals/joss-reviews#2974

If I am not mistaken, the following code blocks are exactly the same. Consider a refactoring.

[numerical/moead.py](https://github.com/moead-framework/framework/blob/master/moead_framework/algorithm/numerical/moead.py#L48)  (48–91)
 [combinatorial/moead.py](https://github.com/moead-framework/framework/blob/master/moead_framework/algorithm/combinatorial/moead.py#L52) (52–95)
