+++ .github/workflows/pythonapp_push.yml
@@ -30,9 +30,8 @@ jobs:
         python3 -m unittest  
     - name: Generate coverage report
       run: |
-        pip install pytest
         pip install pytest-cov
-        pytest --cov=./ --cov-report=xml
+        python3 -m pytest --cov=./ --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
       with:
