+++ .github/workflows/pythonapp_push.yml
@@ -36,7 +36,7 @@ jobs:
       uses: codecov/codecov-action@v1
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
-        files: ./coverage1.xml,./coverage2.xml
+        files: ./coverage.xml
         directory: ./coverage/reports/
         flags: unittests
         env_vars: OS,PYTHON
