### 2021-02-06 de5c0ea
refactoring test : add knapack test & change folder instances
