+++ moead_framework/test/test_combinatorial_algorithm.py
@@ -18,7 +18,7 @@ class AlgorithmsTest(unittest.TestCase):
         self.number_of_evaluations = 100
 
         project_path = os.path.dirname(os.path.abspath(__file__))
-        self.rmnk = Rmnk(instance_file=project_path + '/data/RMNK/Instances/rmnk_0_2_100_1_0.dat')
+        self.rmnk = Rmnk(instance_file=project_path + '/data/instances/rmnk_0_2_100_1_0.dat')
 
         self.number_of_objective = self.rmnk.function_numbers
         self.number_of_weight = 10
