### 2021-02-06 60e5897
fix knapsack test
