### 2021-02-06 a5c25c1
refactoring tools : remove old and useless function, add more unit tests
