### 2021-02-06 7fc80cd
add unit test for solution object
