### 2021-02-07 12c9b0e
refactoring tests : parent_selector, solution, operators
