+++ moead_framework/test/test_solution.py
@@ -1,5 +1,6 @@
 from copy import copy
 
+import numpy as np
 import random
 import os
 import unittest
@@ -12,6 +13,8 @@ class RmnkTest(unittest.TestCase):
 
     def setUp(self):
         """Init"""
+        random.seed(1)
+        np.random.seed(1)
         project_path = os.path.dirname(os.path.abspath(__file__))
         self.problem = Rmnk(project_path + '/data/instances/rmnk_0_2_100_1_0.dat')
         self.solution = self.problem.generate_random_solution()
@@ -19,13 +22,13 @@ class RmnkTest(unittest.TestCase):
     def test_set_item(self):
         """Test set item"""
         # todo
-        self.assertEqual(self.solution[0], -0.5571641596999998)
+        self.assertEqual(self.solution[0], -0.5107130364200001)
         self.solution[0] = -42
         self.assertEqual(self.solution[0], -42)
 
     def test_repr(self):
         """Test repr'"""
-        self.assertEqual(repr(self.solution), "[-0.5278004935999998, -0.4834306098]")
+        self.assertEqual(repr(self.solution), "[-0.5107130364200001, -0.5078202645999998]")
 
     def test_copy(self):
         """"Test the copy of solution"""
