+++ moead_framework/core/genetic_operator/numerical/moead_de_operators.py
@@ -14,10 +14,11 @@ class MoeadDeOperators(GeneticOperator):
         solution2 = self.solutions[1]
         solution3 = self.solutions[2]
 
-        child_cross = DifferentialEvolutionCrossover(solutions=[solution1.solution,
-                                                                solution2.solution,
-                                                                solution3.solution]).run()
-        child_mut = PolynomialMutation(solution=[child_cross])
+        child_cross = DifferentialEvolutionCrossover(solutions=[solution1,
+                                                                solution2,
+                                                                solution3]).run()
+
+        child_mut = PolynomialMutation(solutions=[child_cross]).run()
 
         return child_mut
 
