+++ moead_framework/test/test_genetic_operator.py
@@ -3,7 +3,8 @@ import random
 import numpy as np
 
 from moead_framework.core.genetic_operator.combinatorial import Crossover, BinaryMutation
-from moead_framework.core.genetic_operator.numerical import DifferentialEvolutionCrossover, PolynomialMutation
+from moead_framework.core.genetic_operator.numerical import DifferentialEvolutionCrossover, PolynomialMutation, \
+    MoeadDeOperators
 from moead_framework.problem.numerical import Zdt1
 
 
@@ -61,3 +62,16 @@ class GeneticOperatorsTest(unittest.TestCase):
         self.assertEqual(new_solution, [0, 0.9568466893460874, 0.7769962533153996,
                                         0.3009457924509838, 0.2869152283255434])
 
+    def test_moead_de_operator(self):
+        """Test numerical de crossover"""
+        problem = Zdt1(5)
+        array_solution1 = problem.generate_random_solution().solution
+        array_solution2 = problem.generate_random_solution().solution
+        array_solution3 = problem.generate_random_solution().solution
+
+        new_solution = MoeadDeOperators(solutions=[array_solution1,
+                                                   array_solution2,
+                                                   array_solution3]).run()
+
+        self.assertEqual(new_solution, [0.021382317860195066, 1, 0.7327909561319956,
+                                        0.17080281343876919, 0.2590886275388289])
