### 2021-02-07 03442ee
refactoring abstract class and test for moead-DRA
