### 2021-02-07 moead-framework/framework#12 @geoffreyp
I included codecov in my CI workflow.
On the dev branch the coverage grows from 75% to 96.84%  https://codecov.io/gh/moead-framework/framework/branch/dev
It will be merged for the next release
