+++ .github/workflows/pythonapp_push.yml
@@ -6,13 +6,13 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        # in this example, there is a newer version already installed, 3.7.7, so the older version will be downloaded
+        python-version: ['3.6', '3.7', '3.8', '3.9']
 
     steps:
     - uses: actions/checkout@master
-    - name: Set up Python 3.7
-      uses: actions/setup-python@master
-      with:
-        python-version: 3.7
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
