### 2021-02-07 moead-framework/framework#10 @geoffreyp
Documentation updated, Readme updated on dev branch
