### 2021-02-07 fe01b2d
Create CODE_OF_CONDUCT.md
