### 2021-02-07 moead-framework/framework#13 @geoffreyp
The documentation is updated:
- https://moead-framework.github.io/documentation/html/index.html
- https://moead-framework.github.io/documentation/html/contributing.html

Two new files:
- https://github.com/moead-framework/framework/blob/master/CONTRIBUTING.md
- https://github.com/moead-framework/framework/blob/master/CODE_OF_CONDUCT.md
