+++ docs/index.rst
@@ -1,0 +1,31 @@
+.. moead-framework documentation master file, created by
+   sphinx-quickstart on Sat Nov  2 12:51:33 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to Moead-Framework's documentation!
+===========================================
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   install
+   documentation
+   tuto
+   examples
+   references
+   contributing
+
+
+The **Multi-objective evolutionary algorithm based on decomposition (MOEA/D)** is a general-purpose algorithm 
+for approximating the Pareto set of multi-objective optimization problems. It decomposes the original 
+multi-objective problem into a number of 
+single-objective optimization sub-problems and then uses an evolutionary process to optimize these 
+sub-problems simultaneously and cooperatively. **MOEA/D** is a state-of-art algorithm in aggregation-based 
+approaches for multi-objective optimization.
+
+The goal of the *moead-framework* python package is to provide a modular framework for scientists and 
+researchers interested in experimenting with **MOEA/D** and its numerous variants.
+
+If you have any questions about the framework or the project, feel free to create an issue on github: https://github.com/moead-framework/framework.
