+++ docs/documentation.rst
@@ -1,0 +1,293 @@
+.. _components:
+Components of MOEA/D framework
+===========================================
+
+
+Problems 
+--------------------------------------
+
+========================================= ======================================================= ===================================================================
+Common Name                               Name in the framework                                   Parameters
+========================================= ======================================================= ===================================================================
+rho MNK-Landscapes :cite:`rmnk`           :class:`moead_framework.problem.combinatorial.rmnk`     :code:`instance_file` : path of the instance file
+MUBQP :cite:`mubqp`                       :class:`moead_framework.problem.combinatorial.mubqp`    :code:`instance_file` : path of the instance file
+Multi-objective Knapsack :cite:`mokp`     :class:`moead_framework.problem.combinatorial.knapsack` :code:`objective_number` : number of objective, :code:`instance_file` : path of the instance file
+Zdt1 :cite:`zdt`                           :class:`moead_framework.problem.numerical.zdt`          :code:`size` : size of solutions
+========================================= ======================================================= ===================================================================
+ 
+Four benchmark problems are available in the framework. You can find file instances for combinatorial problems
+in this repository https://github.com/moead-framework/data/tree/master/problem. 
+
+.. note:: You can implement your own problem by following :ref:`this tutoriel<tuto-problem>`.
+
+Three functions are available in a problem object: 
+
+============================================================ ===================================================================
+Function                                                     Comments
+============================================================ ===================================================================
+:class:`problem.generate_random_solution(evaluate)`          Generate a random solution. The parameter :code:`evaluate` is a boolean to specify if the solution have to be evaluated. The default value is :code:`True`.
+:class:`problem.generate_solution(x, evaluate)`              Convert the list :code:`x` to a Solution object. The parameter :code:`evaluate` is a boolean to specify if the solution have to be evaluated. The default value is :code:`True`.
+:class:`problem.f(i, solution)`                              Return the fitness value of the :code:`solution` object for the function :code:`i`.
+============================================================ ===================================================================
+
+A solution is a :code:`OneDimensionSolution` object. This object save the array in the attribute 
+:code:`OneDimensionSolution.solution`, the fitness is saved in the array  :code:`OneDimensionSolution.F` if the  
+parameter :code:`evaluate` is set to True.
+
+Example with the problem rho MNK-Landscapes:
+
+.. code-block:: python
+
+    In [1]: from moead_framework.problem.combinatorial import Rmnk
+
+    # the file is available here: https://github.com/moead-framework/data/blob/master/problem/RMNK/Instances/rmnk_0_2_20_1_0.dat
+    # rmnk instance with parameters rho=0, m=2, n=20, k=1 and a seed of 0
+    In [2]: file_rmnk = "rmnk_0_2_20_1_0.dat"  
+    In [3]: problem = Rmnk(instance_file=file_rmnk) 
+
+    # Generate a random solution
+    In [4]: solution = problem.generate_random_solution()
+
+    # Generate a solution with predefined values
+    In [5]: solution = problem.generate_solution([0,1,1,1,0,1,0,0,1,0])
+
+    # Get objective value of the solution
+    In [6]: array_of_fitness = solution.F
+    In [7]: f0 = solution.F[0]
+    In [8]: f1 = solution.F[1]
+
+
+.. _components_algo:
+Algorithms
+--------------------------------------
+
+====================================================== ================================================================== ===================================================================
+Common Name                                             Name in the framework                                              Comments
+====================================================== ================================================================== ===================================================================
+Original MOEA/D (combinatorial) :cite:`moead`           :class:`moead_framework.algorithm.combinatorial.moead`             The original algorithm for combinatorial optimization
+Original MOEA/D (numerical) :cite:`moead`               :class:`moead_framework.algorithm.numerical.moead`                 The original algorithm for numerical optimization
[…]
