+++ docs/contributing.rst
@@ -1,0 +1,25 @@
+
+Contributing
+===========================================
+
+When contributing to this repository, please first discuss the change you wish to make via issue,
+email, or any other method with the owners of this repository before making a change. 
+
+Please note we have a code of conduct, please follow it in all your interactions with the project.
+
+Pull Request Process
+________________________________
+
+1. Ensure any install or build dependencies are removed before the end of the layer when doing a 
+   build.
+2. Update the README.md with details of changes to the interface, this includes new environment 
+   variables, exposed ports, useful file locations and container parameters.
+3. Update the documentation with a new Pull Request on the repository https://github.com/moead-framework/documentation.
+4. Don't forget to add unit tests.
+5. Create a new branch and a Pull Request to merge your work on the **dev** branch of the project.
+
+
+Code of Conduct
+________________________________
+
+The code of conduct is available here: https://github.com/moead-framework/framework/blob/master/CODE_OF_CONDUCT.md.
