+++ docs/index.rst
@@ -12,6 +12,7 @@ Welcome to Moead-Framework's documentation!
 
    install
    documentation
+   api
    tuto
    examples
    references
