+++ moead_framework/problem/combinatorial/mubqp.py
@@ -4,8 +4,15 @@ from moead_framework.solution.one_dimension_solution import OneDimensionSolution
 
 
 class Mubqp(Problem):
-
+    """
+    Implementation of the Multiobjective Unconstrained Binary Quadratic Programming problem.
+    The problem is compatible with files generated
+    by the mocobench generator http://mocobench.sourceforge.net/index.php?n=Problem.MUBQP
+    """
     def __init__(self, instance_file):
+        """
+        :param instance_file: {string} text file generated by the mubqp generator : http://mocobench.sourceforge.net/index.php?n=Problem.RMNK#Code
+        """
         file = open(instance_file, 'r')
         file_content = list(map(str.strip, file.readlines()))
         file.close()
