+++ docs/moead_framework/moead_framework.problem.combinatorial.knapsack.KnapsackProblem.rst
@@ -1,0 +1,28 @@
+moead\_framework.problem.combinatorial.knapsack.KnapsackProblem
+===============================================================
+
+.. currentmodule:: moead_framework.problem.combinatorial.knapsack
+
+.. autoclass:: KnapsackProblem
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~KnapsackProblem.__init__
+      ~KnapsackProblem.f
+      ~KnapsackProblem.generate_random_solution
+      ~KnapsackProblem.generate_solution
+      ~KnapsackProblem.penality
+      ~KnapsackProblem.profit_of_solution
+      ~KnapsackProblem.weight_of_solution
+   
+   
+
+   
+   
+   
