+++ docs/api.rst
@@ -12,3 +12,25 @@ __________________________________________
    moead_framework.problem.combinatorial.rmnk.Rmnk
    moead_framework.problem.combinatorial.mubqp.Mubqp
    moead_framework.problem.combinatorial.knapsack.KnapsackProblem
+   moead_framework.problem.numerical.zdt.Zdt1
+
+
+Solution
+__________________________________________
+
+.. autosummary::
+   :nosignatures:
+   :toctree: moead_framework
+
+   moead_framework.solution.one_dimension_solution.OneDimensionSolution
+
+
+Algorithms
+__________________________________________
+
+.. autosummary::
+   :nosignatures:
+   :undoc-members:
+   :toctree: moead_framework
+
+   moead_framework.algorithm.combinatorial.moead.Moead
