+++ docs/moead_framework/moead_framework.solution.one_dimension_solution.OneDimensionSolution.rst
@@ -1,0 +1,30 @@
+moead\_framework.solution.one\_dimension\_solution.OneDimensionSolution
+=======================================================================
+
+.. currentmodule:: moead_framework.solution.one_dimension_solution
+
+.. autoclass:: OneDimensionSolution
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~OneDimensionSolution.__init__
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~OneDimensionSolution.F
+      ~OneDimensionSolution.distance
+      ~OneDimensionSolution.solution
+   
+   
