### 2021-02-09 9027640
update documentation API-algorithms
