+++ moead_framework/algorithm/combinatorial/moead_sps_random.py
@@ -1,9 +1,11 @@
-import random
 from moead_framework.algorithm.combinatorial.moead import Moead
 from moead_framework.core.sps_strategy.sps_random_and_boundaries import SpsRandomAndBoundaries
 
 
 class MoeadSPSRandom(Moead):
+    """
+    Implementation of MOEA/D with the component SPS Random.
+    """
 
     def __init__(self, problem,
                  max_evaluation,
@@ -18,6 +20,21 @@ class MoeadSPSRandom(Moead):
                  parent_selector=None,
                  weight_file=None
                  ):
+        """
+        Constructor of the algorithm.
+
+        :param problem: {:class:`~moead_framework.problem.Problem`} problem to optimize
+        :param max_evaluation: {integer} maximum number of evaluation
+        :param number_of_objective: {integer} number of objective in the problem
+        :param number_of_weight: {integer} number of weight vector used to decompose the problem
+        :param number_of_weight_neighborhood: {integer} size of the neighborhood
+        :param number_of_subproblem_to_visit: {integer} number of sub-problems to visit randomly at each generation
+        :param aggregation_function: {:class:`~moead_framework.aggregation.functions.AggregationFunction`}
+        :param genetic_operator: Optional -- {:class:`~moead_framework.core.genetic_operator.abstract_operator.GeneticOperator`} The default operator depends of the problem type (combinatorial / numerical)
+        :param parent_selector: Optional -- {:class:`~moead_framework.core.parent_selector.abstract_parent_selector.ParentSelector`} The default operator depends of the number of solution required by the genetic operator
+        :param mating_pool_selector: Optional -- {:class:`~moead_framework.core.selector.abstract_selector.MatingPoolSelector`} The default selector is {:class:`~moead_framework.core.selector.closest_neighbors_selector.ClosestNeighborsSelector`}
+        :param weight_file: todo need refactoring
+        """
 
         super().__init__(problem,
                          max_evaluation,
