+++ moead_framework/algorithm/combinatorial/moead.py
@@ -5,6 +5,9 @@ from moead_framework.tool.mop import get_non_dominated, is_duplicated
 
 
 class Moead(AbstractMoead):
+    """
+    Implementation of MOEA/D for combinatorial problems.
+    """
 
     def __init__(self, problem,
                  max_evaluation,
