+++ moead_framework/algorithm/combinatorial/moead_delta_nr.py
@@ -6,6 +6,11 @@ from moead_framework.tool.mop import is_duplicated, get_non_dominated
 
 
 class MoeadDeltaNr(Moead):
+    """
+    Implementation of MOEA/D with parameters delta / nr
+
+    H. Li and Q. Zhang. MOEA/D-DE : Multiobjective Optimization Problems With Complicated Pareto Sets, MOEA/D and NSGA-II. IEEE Transactions on Evolutionary Computation, 13(2):284–302, April 2009. doi:10.1109/TEVC.2008.925798
+    """
 
     def __init__(self, problem,
                  max_evaluation,
@@ -17,8 +22,26 @@ class MoeadDeltaNr(Moead):
                  aggregation_function,
                  sps_strategy=None,
                  number_of_crossover_points=2,
+                 mutation_probability=1,
                  parent_selector=None,
                  weight_file=None):
+        """
+        Constructor of the algorithm.
+
+        :param problem: {:class:`~moead_framework.problem.Problem`} problem to optimize
+        :param max_evaluation: {integer} maximum number of evaluation
+        :param number_of_objective: {integer} number of objective in the problem
+        :param number_of_weight: {integer} number of weight vector used to decompose the problem
+        :param number_of_weight_neighborhood: {integer} size of the neighborhood
+        :param delta: {float} probability to use all the population as neighborhood
+        :param number_of_replacement: {integer} maximum number of solutions replaced in the population for each new offspring generated
+        :param aggregation_function: {:class:`~moead_framework.aggregation.functions.AggregationFunction`}
+        :param parent_selector: Optional -- {:class:`~moead_framework.core.parent_selector.abstract_parent_selector.ParentSelector`} The default operator depends of the number of solution required by the genetic operator
+        :param sps_strategy: Optional -- {:class:`~moead_framework.core.sps_strategy.abstract_sps.SpsStrategy`} The default strategy is {:class:`~moead_framework.core.sps_strategy.sps_all.SpsAllSubproblems`}
+        :param number_of_crossover_points: {integer} number of crossover point
+        :param mutation_probability: {integer} probability of mutation used by the genetic operator
+        :param weight_file: todo need refactoring
+        """
 
         mating_pool_selector = DeltaSelector
         self.delta = delta
@@ -34,16 +57,20 @@ class MoeadDeltaNr(Moead):
                          mating_pool_selector=mating_pool_selector,
                          parent_selector=parent_selector,
                          number_of_crossover_points=number_of_crossover_points,
+                         mutation_probability=mutation_probability,
                          sps_strategy=sps_strategy,
                          weight_file=weight_file)
 
     def update_solutions(self, solution, aggregation_function, sub_problem):
         """
-        number_of_replacement is the maximal number of solutions replaced by each child solution
-        to preserve the diversity.
-        :param solution:
-        :param aggregation_function:
-        :param sub_problem:
+        Update solutions of the population and of the external archive ep.
+
+        Integration of the parameter number_of_replacement (the maximal number of solutions replaced by each child solution
+        to preserve the diversity)
+
+        :param solution: {:class:`~moead_framework.solution.one_dimension_solution.OneDimensionSolution`} the candidate solution also called offspring
+        :param aggregation_function: {:class:`~moead_framework.aggregation.functions.AggregationFunction`} Aggregation function used to compare solution in a multi-objective context
+        :param sub_problem: {integer} index of the sub-problem currently visited
         :return:
         """
         c = 0
