+++ moead_framework/core/termination_criteria/abstract_termination_criteria.py
@@ -2,6 +2,9 @@ from abc import ABC, abstractmethod
 
 
 class TerminationCriteria(ABC):
+    """
+    Component used to stop the algorithm
+    """
 
     def __init__(self, algorithm_instance):
         self.algorithm = algorithm_instance
@@ -10,5 +13,6 @@ class TerminationCriteria(ABC):
     def test(self):
         """
         Test if the algorithm has to be stopped
-        :return: Boolean
+
+        :return: {boolean}
         """
