+++ moead_framework/core/sps_strategy/sps_random_and_boundaries.py
@@ -5,6 +5,8 @@ from moead_framework.core.sps_strategy.abstract_sps import SpsStrategy
 
 class SpsRandomAndBoundaries(SpsStrategy):
     """
+    Select randomly lambda sub-problems at each generation with boundaries sub-problems
+
     Pruvost, Geoffrey, et al.
     "On the Combined Impact of Population Size and Sub-problem Selection in MOEA/D."
     European Conference on Evolutionary Computation in Combinatorial Optimization (Part of EvoStar).
@@ -12,8 +14,11 @@ class SpsRandomAndBoundaries(SpsStrategy):
     """
     def get_sub_problems(self):
         """
-        Select one random sub problems in each cluster
-        :return: an array of sub problems
+        Select lambda random sub problems in each cluster
+
+        lambda is represented here by the attribute self.algorithm.number_of_subproblem
+
+        :return: {list<integer>} indexes of sub-problems
         """
         range_list = list(range(self.algorithm.number_of_weight))
         xtrem_index = self.get_xtrem_index()
