+++ moead_framework/core/selector/abstract_selector.py
@@ -2,14 +2,19 @@ from abc import ABC, abstractmethod
 
 
 class MatingPoolSelector(ABC):
+    """
+    select the set of solutions (the neighborhood) used to select parents
+    """
 
     def __init__(self, algorithm_instance):
+        """
+        :param algorithm_instance: {:class:`~moead_framework.algorithm.abstract_moead.py.AbstractMoead`} instance of the algorithm
+        """
         self.algorithm = algorithm_instance
 
     @abstractmethod
     def select(self, sub_problem):
         """
-        select the set of solutions (the neighborhood) before to select parents
-        :param sub_problem:
-        :return: list of index that represents solutions in the population
+        :param sub_problem: {integer} index of the current sub-problem visited
+        :return: {list<integer>}
         """
