+++ moead_framework/aggregation/tchebycheff.py
@@ -3,11 +3,19 @@ from .functions import AggregationFunction
 
 class Tchebycheff(AggregationFunction):
     """
-       Tchebycheff Approach
-       Minimize the maximal distance |f_i(x) - z_i(x)| * w_i
-       """
+    Tchebycheff aggregation function.
+
+    Minimize the maximal distance |f_i(x) - z_i(x)| * w_i
+    """
 
     def is_better(self, old_value, new_value):
+        """
+        Allow to compare 2 aggregations values for the tchebycheff function in a minimization context.
+
+        :param old_value: {float}
+        :param new_value: {float}
+        :return: {boolean} True if new_value is better than old_value.
+        """
         return new_value < old_value
 
     def run(self, solution, number_of_objective, weights, sub_problem, z):
