+++ docs/api.rst
@@ -1,5 +1,6 @@
+.. _components:
 
-API
+Components of MOEA/D framework
 ===========================================
 
 Problems
@@ -14,6 +15,11 @@ __________________________________________
    moead_framework.problem.combinatorial.knapsack.KnapsackProblem
    moead_framework.problem.numerical.zdt.Zdt1
 
+You can find examples of file instances for combinatorial problems
+in this repository https://github.com/moead-framework/data/tree/master/problem.
+
+.. note:: You can implement your own problem by following :ref:`this tutoriel<tuto-problem>`.
+
 
 Solution
 __________________________________________
@@ -27,6 +33,8 @@ __________________________________________
 
 Algorithms
 __________________________________________
+
+These algorithms are implemented on the basis of the original papers (:cite:`moead`, :cite:`moead_de`, :cite:`moead_dra`, :cite:`gpruvost_evocop2020`).
 
 For combinatorial problems
 ---------------------------------
@@ -50,8 +58,15 @@ For numerical problems
    moead_framework.algorithm.numerical.moead.Moead
 
 
+.. note:: You can implement your own algorithm by following :ref:`this tutoriel<tuto-algo>`.
+
+
 Aggregation functions
 __________________________________________
+
+The aggregation function is used in MOEA/D to decompose the multi-objective problem into several mono-objective sub-problems.
+The two main functions used are the Weighted Sum and the Tchebycheff function. In our framework, the aggregation function
+is a required parameter of the algorithm.
 
 .. autosummary::
    :nosignatures:
@@ -63,6 +78,10 @@ __________________________________________
 
 Genetic operators
 __________________________________________
+
+A genetic operator is a component used in genetic algorithms to generate offspring by
+using characteristics of parents solutions. In the framework, these operators are used in the component **offspring_generator**.
+
 
 For combinatorial problems
 ---------------------------------
@@ -91,6 +110,9 @@ For numerical problems
 Offspring generator
 __________________________________________
 
+The offspring generator is the component that manage all the process to generate new offspring by
+using components **Parent Selector** and **Genetic operator** for example.
+
 .. autosummary::
    :nosignatures:
    :toctree: moead_framework
@@ -100,6 +122,9 @@ __________________________________________
 
 Parent Selector
 __________________________________________
+
+The parent selector is the component used to select solutions in the neighborhood before to use genetic
+operators to generate new offspring.
 
 .. autosummary::
    :nosignatures:
@@ -113,6 +138,8 @@ __________________________________________
 Mating Selector
 __________________________________________
 
+The mating selector is the component used to select the set of solutions where we can find parents solutions.
+
 .. autosummary::
    :nosignatures:
    :toctree: moead_framework
@@ -123,6 +150,10 @@ __________________________________________
 
 Sub-problem selection strategy
 __________________________________________
+
+The SPS Strategy :cite:`gpruvost_evocop2020` is the component used to select sub-problems
+(or solutions of the population) that will be visited during the next
+generation of MOEA/D.
 
 .. autosummary::
    :nosignatures:
@@ -135,6 +166,8 @@ __________________________________________
 
 Termination criteria
 __________________________________________
+
+The termination criteria is the component used to determine when the algorithm has to stop.
 
 .. autosummary::
    :nosignatures:
