+++ docs/contributing.rst
@@ -2,7 +2,7 @@
 Contributing
 ===========================================
 
-When contributing to this repository, please first discuss the change you wish to make via issue,
+When contributing to this project, please first discuss the change you wish to make via issue,
 email, or any other method with the owners of this repository before making a change. 
 
 Please note we have a code of conduct, please follow it in all your interactions with the project.
