### 2021-02-09 moead-framework/framework#15 @geoffreyp
Thank you for the tip!
I move the documentation in this repository and I use autodoc with auto-summary.
The documentation is for the moment based on the dev branch at this url https://moead-framework.github.io/framework/
