### 2021-02-10 moead-framework/framework#16 @geoffreyp
Combinatorial problems implemented (ρMNK-Landscapes, Knapsack and MUBQP) are using instance files. These files are generated with generators (noted in the documentation) by using specific parameters for each instance.  I think it will be complicated to implement these generators in our framework to generate exactly the same instance file with the same seed (to compare the results for the same instance between two algorithms with another framework).

Problem classes are strongly inspired by the example code of the generator's authors (for example: http://svn.code.sf.net/p/mocobench/code/trunk/rmnk/generator/java/)
