### 2021-02-10 5997c4c
update documentation (joss review)
