+++ docs/examples.rst
@@ -2,7 +2,7 @@
 Examples
 ===========================================
 
-Optimise Rmnk with MOEA/D
+Optimise mnk-landscapes with MOEA/D
 --------------------------------------
 .. literalinclude:: examples/example_rmnk_moead.py
   :language: python
