+++ docs/api.rst
@@ -1,176 +1,10 @@
 .. _components:
 
-Components of MOEA/D framework
+Library Reference
 ===========================================
 
-Problems
-__________________________________________
+.. toctree::
+	:maxdepth: 2
 
-.. autosummary::
-   :nosignatures:
-   :toctree: moead_framework
-
-   moead_framework.problem.combinatorial.rmnk.Rmnk
-   moead_framework.problem.combinatorial.mubqp.Mubqp
-   moead_framework.problem.combinatorial.knapsack.KnapsackProblem
-   moead_framework.problem.numerical.zdt.Zdt1
-
-You can find examples of file instances for combinatorial problems
-in this repository https://github.com/moead-framework/data/tree/master/problem.
-
-.. note:: You can implement your own problem by following :ref:`this tutoriel<tuto-problem>`.
-
-
-Solution
-__________________________________________
-
-.. autosummary::
-   :nosignatures:
-   :toctree: moead_framework
-
-   moead_framework.solution.one_dimension_solution.OneDimensionSolution
-
-
-Algorithms
-__________________________________________
-
-These algorithms are implemented on the basis of the original papers (:cite:`moead`, :cite:`moead_de`, :cite:`moead_dra`, :cite:`gpruvost_evocop2020`).
-
-For combinatorial problems
----------------------------------
-
-.. autosummary::
-   :nosignatures:
-   :toctree: moead_framework
-
-   moead_framework.algorithm.combinatorial.moead.Moead
-   moead_framework.algorithm.combinatorial.moead_delta_nr.MoeadDeltaNr
-   moead_framework.algorithm.combinatorial.moead_dra.MoeadDRA
-   moead_framework.algorithm.combinatorial.moead_sps_random.MoeadSPSRandom
-
-For numerical problems
----------------------------------
-
-.. autosummary::
-   :nosignatures:
-   :toctree: moead_framework
-
-   moead_framework.algorithm.numerical.moead.Moead
-
-
-.. note:: You can implement your own algorithm by following :ref:`this tutoriel<tuto-algo>`.
-
-
-Aggregation functions
-__________________________________________
-
-The aggregation function is used in MOEA/D to decompose the multi-objective problem into several mono-objective sub-problems.
-The two main functions used are the Weighted Sum and the Tchebycheff function. In our framework, the aggregation function
-is a required parameter of the algorithm.
-
-.. autosummary::
-   :nosignatures:
-   :toctree: moead_framework
-
-   moead_framework.aggregation.tchebycheff.Tchebycheff
-   moead_framework.aggregation.weighted_sum.WeightedSum
-
-
-Genetic operators
-__________________________________________
-
-A genetic operator is a component used in genetic algorithms to generate offspring by
-using characteristics of parents solutions. In the framework, these operators are used in the component **offspring_generator**.
-
-
-For combinatorial problems
----------------------------------
-
-.. autosummary::
-   :nosignatures:
-   :toctree: moead_framework
-
-   moead_framework.core.genetic_operator.combinatorial.mutation.BinaryMutation
-   moead_framework.core.genetic_operator.combinatorial.crossover.Crossover
-   moead_framework.core.genetic_operator.combinatorial.cross_mut.CrossoverAndMutation
-
-
-For numerical problems
----------------------------------
-
-.. autosummary::
-   :nosignatures:
-   :toctree: moead_framework
-
-   moead_framework.core.genetic_operator.numerical.differential_evolution_crossover.DifferentialEvolutionCrossover
-   moead_framework.core.genetic_operator.numerical.polynomial_mutation.PolynomialMutation
-   moead_framework.core.genetic_operator.numerical.moead_de_operators.MoeadDeOperators
-
-
-Offspring generator
-__________________________________________
-
[…]
