### 2021-02-10 moead-framework/framework#14 @geoffreyp
I changed the architecture of the documentation
