### 2021-02-10 2b55994
refactoring duplicated code. Issue #18
