+++ moead_framework/algorithm/abstract_moead.py
@@ -1,5 +1,4 @@
 import numpy as np
-from abc import ABC, abstractmethod
 
 from moead_framework.core.offspring_generator.offspring_generator import OffspringGeneratorGeneric
 from moead_framework.core.selector.closest_neighbors_selector import ClosestNeighborsSelector
@@ -8,7 +7,7 @@ from moead_framework.core.termination_criteria.max_evaluation import MaxEvaluati
 from moead_framework.tool.mop import is_duplicated, get_non_dominated, generate_weight_vectors
 
 
-class AbstractMoead(ABC):
+class AbstractMoead:
 
     def __init__(self, problem, max_evaluation, number_of_objective, number_of_weight, number_of_weight_neighborhood,
                  aggregation_function,
@@ -55,6 +54,8 @@ class AbstractMoead(ABC):
         self.z = self.init_z()
         self.b = self.generate_closest_weight_vectors()
         self.current_sub_problem = -1
+        self.current_eval = 1
+        self.mating_pool = []
 
         if sps_strategy is None:
             self.sps_strategy = SpsAllSubproblems(algorithm_instance=self)
@@ -74,7 +75,6 @@ class AbstractMoead(ABC):
         else:
             self.offspring_generator = offspring_generator(algorithm_instance=self)
 
-    @abstractmethod
     def run(self, checkpoint=None):
         """
         Execute the algorithm.
@@ -82,8 +82,24 @@ class AbstractMoead(ABC):
         :param checkpoint: {function} The default value is None. The checkpoint can be used to save data during the process
         :return:
         """
+        while self.termination_criteria.test():
 
-    @abstractmethod
+            # For each sub-problem i
+            for i in self.get_sub_problems_to_visit():
+
+                if checkpoint is not None:
+                    checkpoint()
+
+                self.update_current_sub_problem(sub_problem=i)
+                self.mating_pool = self.mating_pool_selection(sub_problem=i)[:]
+                y = self.generate_offspring(population=self.mating_pool)
+                y = self.repair(solution=y)
+                self.update_z(solution=y)
+                self.update_solutions(solution=y, aggregation_function=self.aggregation_function, sub_problem=i)
+                self.current_eval += 1
+
+        return self.ep
+
     def update_solutions(self, solution, aggregation_function, sub_problem):
         """
         Update solutions of the population and of the external archive ep
@@ -93,6 +109,25 @@ class AbstractMoead(ABC):
         :param sub_problem: {integer} index of the sub-problem currently visited
         :return:
         """
+        for j in self.b[sub_problem]:
+            y_score = aggregation_function.run(solution=solution,
+                                               number_of_objective=self.number_of_objective,
+                                               weights=self.weights,
+                                               sub_problem=j,
+                                               z=self.z)
+
+            pop_j_score = aggregation_function.run(solution=self.population[j],
+                                                   number_of_objective=self.number_of_objective,
+                                                   weights=self.weights,
+                                                   sub_problem=j,
+                                                   z=self.z)
+
+            if aggregation_function.is_better(pop_j_score, y_score):
+                self.population[j] = solution
+
+                if not is_duplicated(x=solution, population=self.ep, number_of_objective=self.number_of_objective):
+                    self.ep.append(solution)
+                    self.ep = get_non_dominated(self.ep)
 
     def get_sub_problems_to_visit(self):
         """
