+++ moead_framework/algorithm/numerical/moead.py
@@ -2,7 +2,6 @@ from moead_framework.algorithm.abstract_moead import AbstractMoead
 from moead_framework.core.genetic_operator.numerical.moead_de_operators import DifferentialEvolutionCrossover
 from moead_framework.core.parent_selector.two_random_and_current_parent_selector import \
     TwoRandomAndCurrentParentSelector
-from moead_framework.tool.mop import get_non_dominated, is_duplicated
 
 
 class Moead(AbstractMoead):
@@ -22,9 +21,23 @@ class Moead(AbstractMoead):
                  sps_strategy=None,
                  offspring_generator=None,
                  weight_file=None):
+        """
+        Constructor of the algorithm.
 
-        self.current_eval = 1
-        self.mating_pool = []
+        :param problem: {:class:`~moead_framework.problem.Problem`} problem to optimize
+        :param max_evaluation: {integer} maximum number of evaluation
+        :param number_of_objective: {integer} number of objective in the problem
+        :param number_of_weight: {integer} number of weight vector used to decompose the problem
+        :param number_of_weight_neighborhood: {integer} size of the neighborhood
+        :param aggregation_function: {:class:`~moead_framework.aggregation.functions.AggregationFunction`}
+        :param termination_criteria: Optional -- {:class:`~moead_framework.core.termination_criteria.abstract_termination_criteria.TerminationCriteria`} The default component is {:class:`~moead_framework.core.termination_criteria.max_evaluation.MaxEvaluation`}
+        :param genetic_operator: Optional -- {:class:`~moead_framework.core.genetic_operator.abstract_operator.GeneticOperator`} The default operator is :class:`~moead_framework.core.genetic_operator.numerical.differential_evolution_crossover.DifferentialEvolutionCrossover`
+        :param parent_selector: Optional -- {:class:`~moead_framework.core.parent_selector.abstract_parent_selector.ParentSelector`} The default operator is :class:`~moead_framework.core.parent_selector.two_random_and_current_parent_selector.TwoRandomAndCurrentParentSelector`
+        :param mating_pool_selector: Optional -- {:class:`~moead_framework.core.selector.abstract_selector.MatingPoolSelector`} The default selector is {:class:`~moead_framework.core.selector.closest_neighbors_selector.ClosestNeighborsSelector`}
+        :param sps_strategy: Optional -- {:class:`~moead_framework.core.sps_strategy.abstract_sps.SpsStrategy`} The default strategy is {:class:`~moead_framework.core.sps_strategy.sps_all.SpsAllSubproblems`}
+        :param offspring_generator: Optional -- {:class:`~moead_framework.core.offspring_generator.abstract_mating.OffspringGenerator`} The default generator is {:class:`~moead_framework.core.offspring_generator.offspring_generator.OffspringGeneratorGeneric`}
+        :param weight_file: todo need refactoring
+        """
 
         super().__init__(problem,
                          max_evaluation,
@@ -49,46 +62,4 @@ class Moead(AbstractMoead):
             self.parent_selector = TwoRandomAndCurrentParentSelector(algorithm=self)
         else:
             self.parent_selector = parent_selector
-
-    def run(self, checkpoint=None):
-
-        while self.termination_criteria.test():
-
-            # For each sub-problem i
-            for i in self.get_sub_problems_to_visit():
-
-                if checkpoint is not None:
-                    checkpoint()
-
-                self.update_current_sub_problem(sub_problem=i)
-                self.mating_pool = self.mating_pool_selection(sub_problem=i)[:]
-                y = self.generate_offspring(population=self.mating_pool)
-                y = self.repair(solution=y)
-                self.update_z(solution=y)
[…]

[…]
