### 2021-02-10 0172de2
refactoring weights. Issue #17
