+++ moead_framework/test/test_combinatorial_algorithm.py
@@ -58,7 +58,6 @@ class AlgorithmsTest(unittest.TestCase):
         moead = MoeadDeltaNr(problem=self.rmnk,
                              max_evaluation=self.number_of_evaluations,
                              number_of_objective=self.number_of_objective,
-                             number_of_weight=self.number_of_weight,
                              number_of_weight_neighborhood=self.number_of_weight_neighborhood,
                              number_of_crossover_points=self.number_of_crossover_points,
                              aggregation_function=Tchebycheff,
@@ -78,7 +77,6 @@ class AlgorithmsTest(unittest.TestCase):
         moead = MoeadSPSRandom(problem=self.rmnk,
                                max_evaluation=self.number_of_evaluations,
                                number_of_objective=self.number_of_objective,
-                               number_of_weight=self.number_of_weight,
                                aggregation_function=Tchebycheff,
                                number_of_weight_neighborhood=self.number_of_weight_neighborhood,
                                number_of_crossover_points=self.number_of_crossover_points,
@@ -99,7 +97,6 @@ class AlgorithmsTest(unittest.TestCase):
         moead = MoeadDRA(problem=self.rmnk,
                          max_evaluation=500,
                          number_of_objective=self.number_of_objective,
-                         number_of_weight=50,
                          number_of_weight_neighborhood=self.number_of_weight_neighborhood,
                          aggregation_function=Tchebycheff,
                          number_of_crossover_points=self.number_of_crossover_points,
