### 2021-02-10 moead-framework/framework#17 @geoffreyp
Hello, the documentation is updated to explain how to use and generate these files.
I explain the utility of instance files for problems in #16.

We choose to externalize the weight-vector component by using files because there are many ways to generate weight vectors for multi-objective optimization by decomposition and this way allow the researcher to use the same weight-vectors with different algorithms/framework. We provide some example files with the framework: https://github.com/moead-framework/data/blob/master/weights
