### 2021-02-10 #2974 @geoffreyp
Hello @chkoar and @sjvrijn,
I answered your issues and push the new release(0.5.8) of the framework with your comments and tips, thank you!

I move the documentation in the same repository, the new URL is: https://moead-framework.github.io/framework/
