### 2021-02-11 moead-framework/framework#10 @chkoar
As I can see the package requires `numpy`. Although, I cannot see a minimum required version for `numpy`. Does this mean that it can work with any version of `numpy`?
