### 2021-02-11 moead-framework/framework#10 @chkoar
Classifiers for the supported python versions could be added in the `setup.py`
