### 2021-02-11 moead-framework/framework#14 @chkoar
There is no need for the generated documentation to be included in the repository.
