### 2021-02-11 moead-framework/framework#15 @chkoar
The documentation needs a great improvement. For instance:

![image](https://user-images.githubusercontent.com/11897937/107608164-d2ca1600-6c43-11eb-9bf4-22a6e41a51da.png)
