### 2021-02-11 moead-framework/framework#19 @chkoar
[joss-reviews#2974] A note on reproducibility
Related: openjournals/joss-reviews#2974

I cannot see a way of controlling the seed on a class level. I suppose that the intention is to control the seed globally via the built-in or the numpy random modules since you use both of them in the package. Although, this is not documented in the docstrings or/and in the documentation.

Some things to consider:
1. Control the seed on class level by passing a seed number or a random state generator object.  Document it in the docstrings.
2. Control the seed globally using a single entry point. Make a note in the documentation.
