### 2021-02-11 moead-framework/framework#10 @geoffreyp
Documentation updated: https://moead-framework.github.io/framework/html/install.html
setup.py: https://github.com/moead-framework/framework/blob/master/setup.py
requirements.txt: https://github.com/moead-framework/framework/blob/master/requirements.txt
