+++ moead_framework/problem/combinatorial/mubqp.py
@@ -11,6 +11,8 @@ class Mubqp(Problem):
     """
     def __init__(self, instance_file):
         """
+        Constructor of the problem
+
         :param instance_file: {string} text file generated by the mubqp generator : http://mocobench.sourceforge.net/index.php?n=Problem.RMNK#Code
         """
         file = open(instance_file, 'r')
@@ -56,11 +58,17 @@ class Mubqp(Problem):
 
         return x
 
-    def load_qs(self, array):
+    def load_qs(self, file_content):
+        """
+        Load values of the instance file
+
+        :param file_content: {list<float>} content of the instance file
+        :return:
+        """
         line = 0
         for i in range(self.n):
             for j in range(self.n):
-                s = array[line].split("  ")
+                s = file_content[line].split("  ")
                 line += 1
                 for n in range(self.m):
                     self.qs[n][i][j] = int(s[n])
