+++ moead_framework/solution/solution.py
@@ -1,7 +1,5 @@
 class Solution:
 
-    F = []
-
     def __init__(self, solution, f=None):
         if f is None:
             f = []
