+++ moead_framework/core/sps_strategy/abstract_sps.py
@@ -5,6 +5,8 @@ class SpsStrategy(ABC):
 
     def __init__(self, algorithm_instance):
         """
+        Constructor of the Sub-Problem Selection Strategy
+
         :param algorithm_instance: {:class:`~moead_framework.algorithm.abstract_moead.py.AbstractMoead`} instance of the algorithm
         """
         self.algorithm = algorithm_instance
@@ -12,7 +14,7 @@ class SpsStrategy(ABC):
     @abstractmethod
     def get_sub_problems(self):
         """
-        get all sub-problems visited in the next generation
+        Get all sub-problems visited in the next generation
 
         :return: {list<integer>} indexes of sub-problems
         """
