+++ moead_framework/core/termination_criteria/abstract_termination_criteria.py
@@ -7,6 +7,11 @@ class TerminationCriteria(ABC):
     """
 
     def __init__(self, algorithm_instance):
+        """
+        Constructor of the termination criteria
+
+        :param algorithm_instance: {:class:`~moead_framework.algorithm.abstract_moead.py.AbstractMoead`} instance of the algorithm
+        """
         self.algorithm = algorithm_instance
 
     @abstractmethod
@@ -14,5 +19,5 @@ class TerminationCriteria(ABC):
         """
         Test if the algorithm has to be stopped
 
-        :return: {boolean}
+        :return: {boolean} True: Continue the process. False: Stop the process
         """
