+++ moead_framework/core/sps_strategy/sps_random_and_boundaries.py
@@ -30,6 +30,11 @@ class SpsRandomAndBoundaries(SpsStrategy):
         return random_indexes
 
     def get_xtrem_index(self):
+        """
+        get boundaries sub-problems
+
+        :return: {list<integer>} indexes of sub-problems
+        """
         xtrem_index = []
         for i in range(self.algorithm.number_of_weight):
             weight = self.algorithm.weights[i]
