### 2021-02-12 moead-framework/framework#15 @geoffreyp
The API documentation has been improved in the commit 0e45c0eb19e7cde58e747b0723228430d13170fc. All parameters and methods are now explained (type + short description)
