### 2021-02-12 moead-framework/framework#19 @geoffreyp
We currently control the seed globally in our projects based on the framework. I add a note in the documentation:
https://moead-framework.github.io/framework/html/tuto.html#manage-the-reproducibility-of-results
