### 2021-02-14 moead-framework/framework#16 @chkoar
Since the implemented combinatorial problems are coupled with specific type of problems it should be instantiated using the specific file type and version.

```python

file_mapping = {
    "rmnk_0_2_100_1_0": ...
}

class Rmnk:
    def __init__(file="rmnk_0_2_100_1_0"):
         if isinstance(file, str):
             ....

```

That way your avoiding the clutter and give the ability to the user to specify the verion of the problem or to pass the file itself is he/she wants to.
