### 2021-02-14 moead-framework/framework#20 @chkoar
[joss-reviews#2974] Number of objectives
Related: openjournals/joss-reviews#2974

The MOEAD algorithms [require](https://github.com/moead-framework/framework/blob/bc0f0128222dd1897418abacd5bca178be41e66b/moead_framework/algorithm/abstract_moead.py#L14) the number of objectives. Problems also [require](https://github.com/moead-framework/framework/blob/bc0f0128222dd1897418abacd5bca178be41e66b/moead_framework/problem/problem.py#L7) the number of objectives.    However, MOEAD require a problem. Why is that?
