### 2021-02-15 moead-framework/framework#16 @geoffreyp
Sorry but I am not sure to understand, you propose to replace the str attribute `instance_file` (for example "/home/mydir/instances/rmnk_0_2_100_1_0.txt") by the attribute `file` which can be a str (as before with `instance_file`) or directly a file `file=open(instance_file, 'r')`?

I don't understand the dict `file_mapping` but for example for the Rmnk problem, instance files can be named "instance1.txt" or "rmnk1.txt" because parameters are read in the content of the file and not in the file name, so we can't check the filename.
