### 2021-02-15 456c94c
issue #20 remove useless parameter in algorithms
