+++ docs/tuto.rst
@@ -68,8 +68,6 @@ Example for :class:`moead_framework.algorithm.combinatorial.moead` :
               # Mandatory parameters
               problem=rmnk,
               max_evaluation = number_of_evaluations,
-              number_of_objective=number_of_objective,
-              number_of_weight=number_of_weight,
               number_of_weight_neighborhood=number_of_weight_neighborhood,
               weight_file=weight_file,
               aggregation_function=Tchebycheff,
@@ -135,8 +133,6 @@ If you want save all non-dominated solutions (attribute :code:`self.ep` in the a
     moead = Moead(
               problem=rmnk,
               max_evaluation = number_of_evaluations,
-              number_of_objective=number_of_objective,
-              number_of_weight=number_of_weight,
               number_of_weight_neighborhood=number_of_weight_neighborhood,
               weight_file=weight_file,
               aggregation_function=Tchebycheff
