+++ README.md
@@ -54,8 +54,6 @@ The package is available in pypi with a linux environment for python 3.6, 3.7, 3
     ###############################
     moead = Moead(problem=rmnk,
                   max_evaluation=number_of_evaluations,
-                  number_of_objective=number_of_objective,
-                  number_of_weight=number_of_weight,
                   number_of_weight_neighborhood=number_of_weight_neighborhood,
                   weight_file=weight_file,
                   aggregation_function=Tchebycheff,
