### 2021-02-15 moead-framework/framework#20 @geoffreyp
I agree with you, the parameter in algorithms is useless because we have the information with the problem.
It is fixed, thank you
