### 2021-02-19 fb7b165
v0.5.9.1: fix type of mutation probability
