### 2021-02-23 moead-framework/framework#14 @sjvrijn
@geoffreyp Have you considered hosting the documentation on https://readthedocs.org/ instead of as GitHub pages?
That way there is no need to rebuild it manually after every update, and you don't have to keep the generated html/css in your repository.
