+++ .github/workflows/docs.yml
@@ -13,7 +13,7 @@ jobs:
     # Standard drop-in approach that should work for most people.
     - uses: ammaraskar/sphinx-action@master
       with:
-        pre-build-command: "pip uninstall sphinx -y && pip install sphinx==v2.4.4 && pip install sphinxcontrib-bibtex==1.0.0 && pip install sphinx_rtd_theme && pip install numpy >= 1.18"
+        pre-build-command: "pip uninstall sphinx -y && pip install sphinx==v2.4.4 && pip install sphinxcontrib-bibtex==1.0.0 && pip install sphinx_rtd_theme && pip install numpy >= 1.18 && pip install moead-framework"
         build-command: "make html"
         docs-folder: "docs"
     # Publish built docs to gh-pages branch.
