### 2021-02-24 moead-framework/framework#14 @geoffreyp
Hello, I tried readthedocs but after some errors about dependencies, I have just created a GitHub action to rebuild automatically the documentation https://github.com/moead-framework/framework/actions/workflows/docs.yml
Thank you!
