### 2021-03-21 #2974 @sjvrijn
Hi @geoffreyp, are you explicitly waiting for our confirmation on whether issues were sufficiently addressed?

If so, maybe a tip/suggestion for next time: create a separate PR corresponding to each issue. That gives a clear location and interface to view and discuss the changes.
