### 2021-03-21 e5a0ff1
Add docs build badge
A 'docs' build-status badge that links to the documentation is common
with readthedocs-supported documentation. Since the documentation in
this repository is hosted differently, this badge basically provides
the same functionality: a quick way to see if the documentation has
successfully built and a direct link to where it is hosted.
