### 2021-03-21 moead-framework/framework#22 @sjvrijn
[joss-reviews#2974] Examples give errors
Related: openjournals/joss-reviews#2974

The examples listed in the README and the documentation result in errors when I try to run them, both for source and PyPI installs.

Also, I think a quick statement outside of the code-block about the two files that are needed would make it easier for people to get started.

## Reproduction
Tested on Windows 10 with Python 3.6.3

### Preparation
```bash
$ mkdir  ~/src/moead-test && cd ~/src/moead-test
# note that the direct files are available without the github interface at .../raw/... instead of .../blob/...
$ wget https://github.com/moead-framework/data/raw/master/problem/RMNK/Instances/rmnk_0_2_100_1_0.dat
$ wget https://github.com/moead-framework/data/raw/master/weights/SOBOL-2objs-10wei.ws
$ vim docs_example.py  # Copied example from https://moead-framework.github.io/framework/html/examples.html
$ vim readme_example.py  # Copied example from README
$ ls
docs_example.py  readme_example.py  rmnk_0_2_100_1_0.dat  SOBOL-2objs-10wei.ws
$ python3 --version
Python 3.6.3
```

### PyPI installation
```bash
$ python3 -m venv pypi-venv

$ source pypi-venv/Scripts/activate
(pypi-venv) $ python3 -m pip install --upgrade pip moead-framework
<snipped for brevity>
Successfully installed moead-framework-0.5.9.1 numpy-1.19.5 pip-21.0.1

(pypi-venv) $ python3 docs_example.py
Traceback (most recent call last):
  File "docs_example.py", line 24, in <module>
    weight_file = "SOBOL-" + str(rmnk.function_numbers) + "objs-" + str(number_of_weight) + "wei.ws"
AttributeError: 'Rmnk' object has no attribute 'function_numbers'

(pypi-venv) $ python3 readme_example.py
Traceback (most recent call last):
  File "readme_example.py", line 49, in <module>
    + ".txt"
NameError: name 'number_of_crossover_points' is not defined

(pypi-venv) $ deactivate
```

### Source installation
```bash
$ python3 -m venv src-venv

$ source src-venv/Scripts/activate

(src-venv) $ python3 -m pip install --upgrade pip
<snipped for brevity>
Successfully installed pip-21.0.1

(src-venv) $ cd ../moead-framework/ && python3 -m pip install . && cd ../moead-test/
Processing d:\src\moead-framework
Collecting numpy
  Using cached numpy-1.19.5-cp36-cp36m-win_amd64.whl (13.2 MB)
Using legacy 'setup.py install' for moead-framework, since package 'wheel' is not installed.
Installing collected packages: numpy, moead-framework
    Running setup.py install for moead-framework ... done
Successfully installed moead-framework-0.5.9.1 numpy-1.19.5

(src-venv) $ python3 docs_example.py
Traceback (most recent call last):
  File "docs_example.py", line 24, in <module>
    weight_file = "SOBOL-" + str(rmnk.function_numbers) + "objs-" + str(number_of_weight) + "wei.ws"
AttributeError: 'Rmnk' object has no attribute 'function_numbers'

(src-venv) $ python3 readme_example.py
Traceback (most recent call last):
  File "readme_example.py", line 49, in <module>
    + ".txt"
NameError: name 'number_of_crossover_points' is not defined

(src-venv) $ deactivate
```
