### 2021-03-21 moead-framework/framework#19 @sjvrijn
I strongly agree with @chkoar 's 2nd point that a single point of entry would be better. E.g. adding a `set_seed` function to your `moead_framework.tool` subpackage:

```python
import random
import numpy as np

def set_seed(seed):
    random.seed(seed)
    np.random.seed(seed)
```

This would allow you to easily switch random implementations later without bothering the user, for example if you switch everything to numpy using the new recommended [`np.random.Generator` interface](https://numpy.org/doc/stable/reference/random/generator.html).
