### 2021-03-21 moead-framework/framework#23 @sjvrijn
[joss-reviews#2974] How to obtain pareto-optimal solutions from moead?
Related: openjournals/joss-reviews#2974

As demonstrated by the provided examples, the `moead.run()` method returns the found pareto-set of solution evaluations, i.e. the output of `f(function_id, solution)`.
I can't however find which (input) solutions correspond to the returned outputs.

How should these solutions be accessed?
