### 2021-03-21 moead-framework/framework#24 @sjvrijn
[joss-reviews#2974] Paper: further clarify statement of need
Related: openjournals/joss-reviews#2974

Currently, the statement of need centers on other (Python) packages `pymoo`, `pygmo` and `jMetal` not providing a modular interface, and R package `MOEADr` only doing so in R. However, this doesn't answer the question:
> 'If `MOEADr` already exists and provides the desired functionality, why is it not sufficient?'

Some more detailed questions to guide the level of detail for what I think should be in the statement of need:
 - Is there a fundamental problem with the `MOEADr` R package?
 - Or is the main goal to bring it's functionality to Python?
   - If so, what are the (significant) benefits over using `MOEADr` in combination with e.g. [`rpy2`](https://rpy2.github.io/)?
   - Or is there a significant difference in support in the wider ecosystem such as benchmark problem implementations?
 - Does `moead-framework` fundamentally improve on `MOEADr` somehow, e.g. by implementing more modules?
 - ...

Just to note: I much prefer Python over R myself, so I think it's perfectly fine if the desire to program in Python is the main reason, but that should be clearly stated and motivated in the paper. Otherwise, (further) development effort might be better spent on creating/improving a Python interface for an existing implementation such as `MOEADr`.
