### 2021-03-21 moead-framework/framework#25 @sjvrijn
[joss-reviews#2974] Paper: grammar/spelling
Related: openjournals/joss-reviews#2974

I'll use this issue to gather all spelling/grammar issues I find in the paper. Additional remarks will be added as content-related issues are fixed (e.g. #24)

6: either capitalize all or none of the relevant words '[M]ulti-[O]bjective [E]volutionary [A]lgorithm based on [D]ecomposition'
10: state-of-[**the**]-art
19: These ~software~ [implementations]
20: state-of-[**the**]-art ... abstraction[,] but ...
21: do not enable [detailed testing and analysis of the various algorithm's components' behavior.]
23: allows [the definition of] different
37&38: ~his/her~ [their] (suggestion, not required)
