### 2021-03-22 #2974 @geoffreyp
Hello @sjvrijn, my message was just to inform you that you can continue the review because the previous issues have been resolved. I have seen your new issues, I will deal with them as soon as possible.

Thanks for the tip, I will use separate PR now instead of writing the issue number in the commit
