### 2021-03-22 moead-framework/framework#25 @geoffreyp
The PR: https://github.com/moead-framework/framework/pull/26/files
The new pdf is generated here : https://github.com/moead-framework/framework/suites/2316652138/artifacts/48675775
