### 2021-03-23 0067a2e
issue #22: fix examples
