+++ example_rmnk_moead.py
@@ -7,21 +7,20 @@ from moead_framework.tool.result import save_population
 ###############################
 #   Initialize the problem    #
 ###############################
+# The file is available here : https://github.com/moead-framework/data/blob/master/problem/RMNK/Instances/rmnk_0_2_100_1_0.dat
 # Others instances are available here : https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances
 instance_file = "moead_framework/test/data/instances/rmnk_0_2_100_1_0.dat"
 rmnk = Rmnk(instance_file=instance_file)
-
 
 #####################################
 #      Initialize the algorithm     #
 #####################################
 number_of_weight = 10
 number_of_weight_neighborhood = 2
-number_of_crossover_points = 4
 number_of_evaluations = 1000
+# The file is available here : https://github.com/moead-framework/data/blob/master/weights/SOBOL-2objs-10wei.ws
 # Others weights files are available here : https://github.com/moead-framework/data/tree/master/weights
 weight_file = "moead_framework/test/data/weights/SOBOL-" + str(rmnk.number_of_objective) + "objs-" + str(number_of_weight) + "wei.ws"
-
 
 
 ###############################
@@ -30,13 +29,11 @@ weight_file = "moead_framework/test/data/weights/SOBOL-" + str(rmnk.number_of_ob
 moead = Moead(problem=rmnk,
               max_evaluation=number_of_evaluations,
               number_of_weight_neighborhood=number_of_weight_neighborhood,
-              number_of_crossover_points=number_of_crossover_points,
               weight_file=weight_file,
               aggregation_function=Tchebycheff,
               )
 
 population = moead.run()
-
 
 ###############################
 #       Save the result       #
@@ -44,7 +41,6 @@ population = moead.run()
 save_file = "moead-rmnk" + str(rmnk.number_of_objective) \
             + "-N" + str(number_of_weight) \
             + "-T" + str(number_of_weight_neighborhood) \
-            + "-CP" + str(number_of_crossover_points) \
             + "-iter" + str(number_of_evaluations) \
             + ".txt"
 
