+++ README.md
@@ -23,6 +23,17 @@ The package is available in pypi with a linux environment for python 3.6, 3.7, 3
     
 # Example
 
+The example requires two files : 
+
+- ```instance_file``` is required by the [problem](https://moead-framework.github.io/framework/html/moead_framework/moead_framework.problem.combinatorial.rmnk.Rmnk.html#moead_framework.problem.combinatorial.rmnk.Rmnk). The file is available in the framework 
+  in "moead_framework/test/data/instances/" or can be downloaded 
+  from [https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances](https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances)
+  
+- ```weight_file``` is required by the [algorithm](https://moead-framework.github.io/framework/html/moead_framework/moead_framework.algorithm.combinatorial.moead.Moead.html#moead_framework.algorithm.combinatorial.moead.Moead). The file is available in the framework 
+  in "moead_framework/test/data/weights/" or can be downloaded 
+  from [https://github.com/moead-framework/data/tree/master/weights](https://github.com/moead-framework/data/tree/master/weights)
+  
+
     from moead_framework.aggregation import Tchebycheff
     from moead_framework.algorithm.combinatorial import Moead
     from moead_framework.problem.combinatorial import Rmnk
@@ -34,20 +45,19 @@ The package is available in pypi with a linux environment for python 3.6, 3.7, 3
     ###############################
     # The file is available here : https://github.com/moead-framework/data/blob/master/problem/RMNK/Instances/rmnk_0_2_100_1_0.dat
     # Others instances are available here : https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances
-    instance_file = "rmnk_0_2_100_1_0.dat"
+    instance_file = "moead_framework/test/data/instances/rmnk_0_2_100_1_0.dat"
     rmnk = Rmnk(instance_file=instance_file)
     
     
     #####################################
     #      Initialize the algorithm     #
     #####################################
-    number_of_objective = rmnk.number_of_objective
     number_of_weight = 10
-    number_of_weight_neighborhood = 20
+    number_of_weight_neighborhood = 2
     number_of_evaluations = 1000
     # The file is available here : https://github.com/moead-framework/data/blob/master/weights/SOBOL-2objs-10wei.ws
     # Others weights files are available here : https://github.com/moead-framework/data/tree/master/weights
-    weight_file = "SOBOL-" + str(number_of_objective) + "objs-" + str(number_of_weight) + "wei.ws"
+    weight_file = "moead_framework/test/data/weights/SOBOL-" + str(rmnk.number_of_objective) + "objs-" + str(number_of_weight) + "wei.ws"
     
     
     ###############################
@@ -66,10 +76,9 @@ The package is available in pypi with a linux environment for python 3.6, 3.7, 3
     ###############################
     #       Save the result       #
     ###############################
-    save_file = "moead-rmnk" + str(number_of_objective) \
+    save_file = "moead-rmnk" + str(rmnk.number_of_objective) \
                 + "-N" + str(number_of_weight) \
                 + "-T" + str(number_of_weight_neighborhood) \
-                + "-CP" + str(number_of_crossover_points) \
                 + "-iter" + str(number_of_evaluations) \
                 + ".txt"
     
