### 2021-03-23 moead-framework/framework#23 @geoffreyp
The `moead.run()` returns a list of [Solution](https://moead-framework.github.io/framework/html/moead_framework/moead_framework.solution.one_dimension_solution.OneDimensionSolution.html#moead_framework.solution.one_dimension_solution.OneDimensionSolution) and not only the output of `f(function_id, solution)`.
So for each [Solution](https://moead-framework.github.io/framework/html/moead_framework/moead_framework.solution.one_dimension_solution.OneDimensionSolution.html#moead_framework.solution.one_dimension_solution.OneDimensionSolution) x in the list, objective values are available in x.F and variables in x.solution.

I edited the documentation to make it more understandable in the PR https://github.com/moead-framework/framework/pull/28
