### 2021-03-23 3199e15
add the set_seed() function
