+++ docs/tuto.rst
@@ -98,18 +98,15 @@ Reproducibility of results is a major principle for scientific research.
 The feature used here is not specific to the framework but
 can be used for every python project that uses the random and numpy modules.
 
-Because the framework uses the random and numpy modules, you can be sure
-to have the same results by running the same script several times if you
-add the following instructions before initializing problems or algorithms:
+You can be sure to have the same results by running the same script several times if you
+call the ``set_seed(seed)`` method before initializing problems and algorithms:
 
 .. code-block:: python
 
-    import random
-    import numpy
+    from moead_framework.tool.result import set_seed
 
     seed = 0
-    random.seed(seed)
-    np.random.seed(seed)
+    set_seed(seed)
 
 
 You can find more information with the following links:
