### 2021-03-23 moead-framework/framework#19 @geoffreyp
The function is added with the PR https://github.com/moead-framework/framework/pull/29
