### 2021-03-24 moead-framework/framework#24 @geoffreyp
The paper is updated with the PR https://github.com/moead-framework/framework/pull/30
