### 2021-03-27 moead-framework/framework#31 @sjvrijn
Confusing OneDimensionalSolution.__repr__
Related: openjournals/joss-reviews#2974 and #23

The `__repr__` of OneDimensionalSolution currently only shows the objective values without any indication it is a more complex object.

This does not follow [Python's recommended usage of the `__repr__` method](https://docs.python.org/3/reference/datamodel.html#object.__repr__) and goes against the [Zen of Python](https://www.python.org/dev/peps/pep-0020/)'s line:
> Explicit is better than implicit

I would strongly recommend changing it such that it can be round-tripped using `eval(repr(x))`  for any `x = OneDimensionalSolution(...)`
