### 2021-03-27 moead-framework/framework#32 @sjvrijn
[Documentation] Include example of how to use results
Related: openjournals/joss-reviews#2974 and #23

The `Solution`-class used in this framework as a default output is not so intuitive/standard that it can be included unexplained.
A (simple) example of how to extract and use each type of returned/stored data for typical uses would be very helpful, so something like
 - plotting a typical 2D pareto-front
 - obtaining the particular `solution` for some `f` (to use your nomenclature)
 - using the included distance attribute

Using https://documentation.divio.com/ terminology, ideally you would also include some explanation in a 'tutorial', but I think that's more optional than adding the list above as a set of 'how-to's.
