### 2021-03-28 moead-framework/framework#13 @sjvrijn
The `CONTRIBUTING.md` only really covers pull requests and is not referenced anywhere.

My proposal would be:
 - Add a short section to the README, also referring to the `CONTRIBUTING.md` and code of conduct, stating explicitly what someone can do who
   - has a question
   - wants to open an issue
   - wants to contribute by submitting a PR
