### 2021-03-28 29b8589
Fix typo 'tutorial' in main_components.rst
