+++ docs/main_components.rst
@@ -18,7 +18,7 @@ __________________________________________
 You can find examples of file instances for combinatorial problems
 in this repository https://github.com/moead-framework/data/tree/master/problem.
 
-.. note:: You can implement your own problem by following :ref:`this tutoriel<tuto-problem>`.
+.. note:: You can implement your own problem by following :ref:`this tutorial<tuto-problem>`.
 
 
 Solution
@@ -58,7 +58,7 @@ For numerical problems
    moead_framework.algorithm.numerical.moead.Moead
 
 
-.. note:: You can implement your own algorithm by following :ref:`this tutoriel<tuto-algo>`.
+.. note:: You can implement your own algorithm by following :ref:`this tutorial<tuto-algo>`.
 
 
 Aggregation functions
