+++ docs/tuto.rst
@@ -6,13 +6,13 @@ Tutorials
 .. _tuto-problem:
 Implement your own problem
 --------------------------------------
-To make your problem compliant with the framework, your problem class have to extend the class :class:`moead_framework.problem.Problem` 
-and also implement the 3 required functions : 
+To make your problem compliant with the framework, your problem class has to extend the class :class:`moead_framework.problem.Problem`
+and implement the following 3 required functions :
 
-- the fitness function :code:`f(function_id, solution)` has 2 required parameters. The function must returns the objectif value of the solution 
+- the fitness function :code:`f(function_id, solution)` has 2 required parameters. The function must returns the objective value of the solution
   for the objective function_id in parameter.
 
-  .. note:: For a better compatibility with components, problems should be converted in minimization problems.
+  .. note:: For compatibility with existing components, problems should be converted to minimization problems.
 
 .. code-block:: python
     
@@ -22,12 +22,12 @@ and also implement the 3 required functions :
 
 
 
-- The function :code:`generate_solution(array, evaluate)` allows to generate a solution. A solution is an object OneDimensionSolution that contains the representation of the solution in the attribute solution and all fitness values in the F attribute. 
+- The function :code:`generate_solution(array, evaluate)` generates a solution. A solution is a :class:`moead_framework.solution.one_dimensional_dolution.OneDimensionSolution` object that contains the representation of the solution in the `solution` attribute, and all fitness values in the `F` attribute.
  
- .. note:: All components of the framework available in this version are only compatible with OneDimensionSolution. If you want manage new type of solutions, you have to adapt other components of the framework. 
+ .. note:: All components of the framework available in this version are only compatible with OneDimensionSolution. If you want to manage new types of solutions, you have to adapt other components of the framework.
  
- The function :code:`generate_solution(array, evaluate)` takes in parameter the representation of the function and the boolean :code:`evaluate` 
- that determine if the solution save the fitness, the default value must be "True".
+ The function :code:`generate_solution(array, evaluate)` takes as parameter the representation of the function and the boolean :code:`evaluate`
+ that determine if the solution object should save the fitness, the default value must be "True".
 
 .. code-block:: python
     
@@ -43,7 +43,7 @@ and also implement the 3 required functions :
             return x
   
 
-- The function :code:`generate_random_solution(evaluate)` generates a solution **randomly** in the same way than the previous function.
+- The function :code:`generate_random_solution(evaluate)` generates a solution **randomly** in the same way as the previous function.
 
 .. code-block:: python
 
@@ -57,8 +57,8 @@ Implement your own algorithm
 --------------------------------------------------------------------
 
 All components are set with default values to implement the first version of **MOEA/D**.
-You can customize each algorithms of the framework with your own
-components that you can set in parameter of the algorithm contructor.
+You can customize each algorithm in the framework with your own
+components that you can set as parameter of the algorithm contructor.
 
 Example for :class:`moead_framework.algorithm.combinatorial.moead` :
 
@@ -81,14 +81,14 @@ Example for :class:`moead_framework.algorithm.combinatorial.moead` :
               sps_strategy=SpsAllSubproblems,
               offspring_generator=OffspringGeneratorGeneric
               )
-    
-    
 
-If you want manage the way to use all this components, you have to create
+
+
+If you want to manage the way to use all these components, you have to create
 a new algorithm by extending an available algorithm. Examples are available in this repository : https://github.com/moead-framework/framework/tree/master/moead_framework/algorithm.
 
[…]
