### 2021-03-28 moead-framework/framework#33 @sjvrijn
[Documentation] irrelevant __init__ instruction on documentation for e.g. aggregation functions
Related: openjournals/joss-reviews#2974

For example [here](https://moead-framework.github.io/framework/html/moead_framework/moead_framework.aggregation.tchebycheff.Tchebycheff.html) is the instruction
>__init__()
>   Initialize self. See help(type(self)) for accurate signature.

which gives the following unrelated result in a terminal:

```python
>>> from moead_framework.aggregation import Tchebycheff
>>> help(type(Tchebycheff))
Help on class ABCMeta in module abc:

class ABCMeta(builtins.type)
[...]
```
