### 2021-03-28 moead-framework/framework#34 @sjvrijn
Add development requirements
Related: openjournals/joss-reviews#2974

Currently the required packages for building the documentation are only manually specified in the github actions docs.yml file.
A simple `requirements-dev.txt` with those packages (and corresponding install instruction) would help potential contributors to work on the project locally.
