### 2021-03-28 moead-framework/framework#35 @sjvrijn
Disambiguate usage of the word 'generate' for Problems
Related: openjournals/joss-reviews#2974

The word 'generate' is currently used with different meanings.

## Definitions

Let me first clarify how I typically interpret certain words in the context of evolutionary optimization:
 - **Generation**: to newly create some candidate or feasible solution **x** according to some procedure, such as a new random sample
 - **Evaluation**: to determine/calculate the fitness _f_(**x**) according to one or more objectives _f_

## My confusion

For example in the [Problem tutorial](https://moead-framework.github.io/framework/html/tuto.html#implement-your-own-problem) where it introduces the example function as:
> The function generate_solution(array, evaluate) allows to generate a solution.

To me, this function creates a `Solution` object, containing **x** and additionally the **evaluation** _f_(**x**) if `evaluate=True` is given, but no **x** is **generated**, since the **x** is given as the `array` parameter, which has to have been **generated** beforehand.

However, the `generate_random_solution(evaluate)` method listed further on does **generate** a new **x**, in addition to performing the other two steps.

## A potential solution

I think it would be clearest if it were refactored to something like this:

```python
class Problem():
    def __init__(self, ...):
        ... # unchanged from current implementation

    def f(self, function_id, solution):
        ... # unchanged from current implementation

    def evaluate(self, solution):
        """evaluate a feasible solution on all objectives"""
        if not isinstance(solution, OneDimensionSolution):
            solution = OneDimensionSolution(solution)
        solution.F = [self.f(j, solution) for j in range(self.number_of_objective)]
        return solution

    def generate_random_solution(self):
        """create new random feasible solution according to problem domain"""
        ...  # only create the new feasible solution x
        return OneDimensionSolution(x)
```

This way, each of the methods `evaluate` and `generate_random_solution` does exactly one thing without 'overloading' the word 'generate'. Importantly, the encapsulation of **x** in a `OneDimensionSolution` object should either be  automatic so a user does not have to think about it, or should be so explicit (if e.g. the user wants to use a different `Solution class`) that it does not belong inside the definition of a `Problem`.

In the latter case, `solution = OneDimensionSolution(solution)` should be replaced with a `raise ValueError('evaluate() expects a Solution-derived object as input')` and `generate_random_solution()` should just `return x`.

## Discussion

I obviously don't know how large the ripple effects of such a refactor would be throughout the rest of the codebase, but I feel this is the most long-term solution to the confusion I encountered regarding the use of the word 'generate'.

I don't want to be too pedantic about this, so @geoffreyp and @chkoar please let me know if anything is unclear, what you think of this suggestion and if you think this could more easily be cleared up by clarifying relevant documentation instead.
