### 2021-03-28 f412e5b
Fix some typos and layout in tuto.rst
