### 2021-03-28 moead-framework/framework#37 @sjvrijn
[Documentation] Problems considered 'main component' of MOEA/D algorithm?
Related: openjournals/joss-reviews#2974

The included (example) problems are currently listed under 'main components' in the [library reference](https://moead-framework.github.io/framework/html/main_components.html#problems) section of the documentation.
While a problem is a required _argument_ for creating a `MOEAD` algorithm object, I do not feel that a `Problem` is a 'component' of the algorithm in the true sense of that word.
