### 2021-03-28 moead-framework/framework#37 @sjvrijn
If you do not like the proposed solution in PR #38, feel free to discuss why here
