### 2021-03-29 moead-framework/framework#33 @sjvrijn
Also related to #15
