### 2021-03-29 moead-framework/framework#40 @sjvrijn
Remove docs/html folder from master branch?
Related: openjournals/joss-reviews#2974

With the documentation being built on a separate gh-pages branch, does that mean the docs/html folder in master is no longer needed?
If so, the repository can be cleaned up by removing it.
