### 2021-03-29 moead-framework/framework#25 @sjvrijn
One further remark about the following sentence:
>With the moead-framework package, we aim at bringing the modularity of the MOEADr
package by using the flexibility of Python in order to allow the user to update the behavior
of MOEA/D components in their research works, and to propose new variants without being
limited by the software.

When using the verb 'bringing' in this situation, it is usually used as 'bringing ... to ...', meaning I would expect to read something like
> We aim at bringing the modularity of `MOEADr` **to ...**

It is also quite long, so please rewrite it.
