### 2021-03-29 moead-framework/framework#15 @sjvrijn
API documentation is missing for the [solution](https://github.com/moead-framework/framework/tree/master/moead_framework/solution) and [tool](https://github.com/moead-framework/framework/tree/master/moead_framework/tool) folders
