### 2021-03-29 moead-framework/framework#42 @chkoar
[joss-reviews#2974] Ztd1 unfriendly error message
Related: openjournals/joss-reviews#2974

Ztd1 raises `ZeroDivisionError`. It should reaise a more user friendly message with the correct type of the exception.

```python
from moead_framework.problem.numerical import Zdt1
problem = Zdt1(1)
problem.generate_random_solution()
```
