### 2021-03-29 moead-framework/framework#43 @chkoar
[joss-reviews#2974] Empty initializer
Related: openjournals/joss-reviews#2974

Is there a need for [this](https://github.com/moead-framework/framework/blob/892a704f9ce43945d3b2c6343bedee37638fc31e/moead_framework/aggregation/functions.py#L6) empty initializer?
