### 2021-03-29 moead-framework/framework#44 @chkoar
[joss-reviews#2974] Naming confusion
Related: openjournals/joss-reviews#2974

Something feels strange here. There is a package that is called solution that contains a module which is called solution that contains a class which is called solution that requires a parameter that is called solution.
