### 2021-03-29 moead-framework/framework#45 @chkoar
[joss-reviews#2974] Annotation errors
Related: openjournals/joss-reviews#2974

As I can see type annotations are not used. For some code there are type annotations though. Here are some of the errors found.

```python
moead_framework/solution/one_dimension_solution.py:9: error: Need type annotation for 'solution' (hint: "solution: List[<type>] = ...")
moead_framework/solution/one_dimension_solution.py:10: error: Need type annotation for 'F' (hint: "F: List[<type>] = ...")
moead_framework/tool/result.py:67: error: Bracketed expression "[...]" is not valid as a type
moead_framework/tool/result.py:67: note: Did you mean "List[...]"?
moead_framework/tool/result.py:69: error: Incompatible types in assignment (expression has type "List[<nothing>]", variable has type "MultiList")
moead_framework/tool/result.py:71: error: Function "numpy.array" is not valid as a type
moead_framework/tool/result.py:71: note: Perhaps you need "Callable[...]" or a callback protocol?
moead_framework/tool/result.py:86: error: np.array? has no attribute "__iter__" (not iterable)
```
