### 2021-03-29 moead-framework/framework#16 @sjvrijn
Maybe something like [Pooch](https://pypi.org/project/pooch/) can be of help? Although that's more aimed at larger files.
