### 2021-03-30 moead-framework/framework#45 @chkoar
`mypy`
