### 2021-03-30 7c9586a
fix compatibility with numpy < 1.20
