### 2021-03-30 a70c367
remove useless __init__  -  issue #43
