+++ docs/moead_framework/moead_framework.core.genetic_operator.numerical.polynomial_mutation.PolynomialMutation.rst
@@ -1,27 +1,0 @@
-moead\_framework.core.genetic\_operator.numerical.polynomial\_mutation.PolynomialMutation
-=========================================================================================
-
-.. currentmodule:: moead_framework.core.genetic_operator.numerical.polynomial_mutation
-
-.. autoclass:: PolynomialMutation
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~PolynomialMutation.__init__
-      ~PolynomialMutation.mutation
-      ~PolynomialMutation.number_of_solution_is_correct
-      ~PolynomialMutation.repair
-      ~PolynomialMutation.run
-      ~PolynomialMutation.sigma
-   
-   
-
-   
-   
-   
