+++ docs/moead_framework/moead_framework.core.selector.delta_selector.DeltaSelector.rst
@@ -1,23 +1,0 @@
-moead\_framework.core.selector.delta\_selector.DeltaSelector
-============================================================
-
-.. currentmodule:: moead_framework.core.selector.delta_selector
-
-.. autoclass:: DeltaSelector
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~DeltaSelector.__init__
-      ~DeltaSelector.select
-   
-   
-
-   
-   
-   
