+++ docs/moead_framework/moead_framework.algorithm.combinatorial.moead_sps_random.MoeadSPSRandom.rst
@@ -1,33 +1,0 @@
-moead\_framework.algorithm.combinatorial.moead\_sps\_random.MoeadSPSRandom
-==========================================================================
-
-.. currentmodule:: moead_framework.algorithm.combinatorial.moead_sps_random
-
-.. autoclass:: MoeadSPSRandom
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~MoeadSPSRandom.__init__
-      ~MoeadSPSRandom.generate_closest_weight_vectors
-      ~MoeadSPSRandom.generate_offspring
-      ~MoeadSPSRandom.get_sub_problems_to_visit
-      ~MoeadSPSRandom.init_z
-      ~MoeadSPSRandom.initial_population
-      ~MoeadSPSRandom.mating_pool_selection
-      ~MoeadSPSRandom.repair
-      ~MoeadSPSRandom.run
-      ~MoeadSPSRandom.update_current_sub_problem
-      ~MoeadSPSRandom.update_solutions
-      ~MoeadSPSRandom.update_z
-   
-   
-
-   
-   
-   
