+++ docs/moead_framework/moead_framework.algorithm.combinatorial.moead_delta_nr.MoeadDeltaNr.rst
@@ -1,33 +1,0 @@
-moead\_framework.algorithm.combinatorial.moead\_delta\_nr.MoeadDeltaNr
-======================================================================
-
-.. currentmodule:: moead_framework.algorithm.combinatorial.moead_delta_nr
-
-.. autoclass:: MoeadDeltaNr
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~MoeadDeltaNr.__init__
-      ~MoeadDeltaNr.generate_closest_weight_vectors
-      ~MoeadDeltaNr.generate_offspring
-      ~MoeadDeltaNr.get_sub_problems_to_visit
-      ~MoeadDeltaNr.init_z
-      ~MoeadDeltaNr.initial_population
-      ~MoeadDeltaNr.mating_pool_selection
-      ~MoeadDeltaNr.repair
-      ~MoeadDeltaNr.run
-      ~MoeadDeltaNr.update_current_sub_problem
-      ~MoeadDeltaNr.update_solutions
-      ~MoeadDeltaNr.update_z
-   
-   
-
-   
-   
-   
