+++ docs/moead_framework/moead_framework.aggregation.tchebycheff.Tchebycheff.rst
@@ -1,24 +1,0 @@
-moead\_framework.aggregation.tchebycheff.Tchebycheff
-====================================================
-
-.. currentmodule:: moead_framework.aggregation.tchebycheff
-
-.. autoclass:: Tchebycheff
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~Tchebycheff.__init__
-      ~Tchebycheff.is_better
-      ~Tchebycheff.run
-   
-   
-
-   
-   
-   
