### 2021-03-30 moead-framework/framework#40 @geoffreyp
the documentation is cleaned in the PR https://github.com/moead-framework/framework/pull/48
