### 2021-03-30 63a164b
update the __repr__ method of the solution
