### 2021-03-30 moead-framework/framework#31 @geoffreyp
I add more information in the PR https://github.com/moead-framework/framework/pull/49

Before: [..., ...., ....]
Now: Solution(F(x)=[..., ...., ....] ; x=[..,..,..,..,..,..])
