### 2021-03-30 moead-framework/framework#44 @geoffreyp
I aggree with you, there are several possibilities but I don't know which one would be the best?
- rename the `solution.py` to `abstract_solution.py` and the class `Solution` to `AbstractSolution` ?
- delete the `solution.py`file, is it necessary to have an abstract class here ?

And for the attribute (and parameter) `solution`, I've been wanting to rename it for some time to `decision_vector` or `x`?
