### 2021-03-30 moead-framework/framework#34 @geoffreyp
The file `requirements-dev.txt` and the readme is updated in the PR https://github.com/moead-framework/framework/pull/50
