### 2021-03-30 82ce4d6
issue #42 : raise exception in zdt1 if size <= 1
