### 2021-03-30 69afbd9
issue #42 : raise exception in zdt1 if size <= 1
