### 2021-03-30 moead-framework/framework#42 @geoffreyp
It is fixed in the PR https://github.com/moead-framework/framework/pull/51
