### 2021-03-31 d099d4a
update __repre__ and add __str__ for OneDimensionSolution
