### 2021-03-31 3485349
update __repr__ of OneDimensionSolution
