+++ README.md
@@ -33,57 +33,57 @@ The example requires two files :
   in "moead_framework/test/data/weights/" or can be downloaded 
   from [https://github.com/moead-framework/data/tree/master/weights](https://github.com/moead-framework/data/tree/master/weights)
   
-
-    from moead_framework.aggregation import Tchebycheff
-    from moead_framework.algorithm.combinatorial import Moead
-    from moead_framework.problem.combinatorial import Rmnk
-    from moead_framework.tool.result import save_population
+```python
+from moead_framework.aggregation import Tchebycheff   
+from moead_framework.algorithm.combinatorial import Moead   
+from moead_framework.problem.combinatorial import Rmnk  
+from moead_framework.tool.result import save_population
     
     
-    ###############################
-    #   Initialize the problem    #
-    ###############################
-    # The file is available here : https://github.com/moead-framework/data/blob/master/problem/RMNK/Instances/rmnk_0_2_100_1_0.dat
-    # Others instances are available here : https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances
-    instance_file = "moead_framework/test/data/instances/rmnk_0_2_100_1_0.dat"
-    rmnk = Rmnk(instance_file=instance_file)
+###############################
+#   Initialize the problem    #
+###############################
+# The file is available here : https://github.com/moead-framework/data/blob/master/problem/RMNK/Instances/rmnk_0_2_100_1_0.dat
+# Others instances are available here : https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances
+instance_file = "rmnk_0_2_100_1_0.dat"
+rmnk = Rmnk(instance_file=instance_file)
     
     
-    #####################################
-    #      Initialize the algorithm     #
-    #####################################
-    number_of_weight = 10
-    number_of_weight_neighborhood = 2
-    number_of_evaluations = 1000
-    # The file is available here : https://github.com/moead-framework/data/blob/master/weights/SOBOL-2objs-10wei.ws
-    # Others weights files are available here : https://github.com/moead-framework/data/tree/master/weights
-    weight_file = "moead_framework/test/data/weights/SOBOL-" + str(rmnk.number_of_objective) + "objs-" + str(number_of_weight) + "wei.ws"
+#####################################
+#      Initialize the algorithm     #
+#####################################
+number_of_weight = 10
+number_of_weight_neighborhood = 2
+number_of_evaluations = 1000
+# The file is available here : https://github.com/moead-framework/data/blob/master/weights/SOBOL-2objs-10wei.ws
+# Others weights files are available here : https://github.com/moead-framework/data/tree/master/weights
+weight_file = "SOBOL-" + str(rmnk.number_of_objective) + "objs-" + str(number_of_weight) + "wei.ws"
     
     
-    ###############################
-    #    Execute the algorithm    #
-    ###############################
-    moead = Moead(problem=rmnk,
-                  max_evaluation=number_of_evaluations,
-                  number_of_weight_neighborhood=number_of_weight_neighborhood,
-                  weight_file=weight_file,
-                  aggregation_function=Tchebycheff,
-                  )
+###############################
+#    Execute the algorithm    #
+###############################
+moead = Moead(problem=rmnk,
+                max_evaluation=number_of_evaluations,
+                number_of_weight_neighborhood=number_of_weight_neighborhood,
+                weight_file=weight_file,
+                aggregation_function=Tchebycheff,
+                )
     
-    population = moead.run()
+population = moead.run()
     
     
-    ###############################
-    #       Save the result       #
-    ###############################
-    save_file = "moead-rmnk" + str(rmnk.number_of_objective) \
+###############################
+#       Save the result       #
+###############################
[…]
