+++ CONTRIBUTING.md
@@ -3,13 +3,29 @@
 When contributing to this repository, please first discuss the change you wish to make via issue,
 email, or any other method with the owners of this repository before making a change. 
 
-Please note we have a code of conduct, please follow it in all your interactions with the project.
+Please note we have [a code of conduct](https://github.com/moead-framework/framework/blob/master/CODE_OF_CONDUCT.md), please follow it in all your interactions with the project.
 
-## Pull Request Process
+
+## You want to share a bug or an idea ?
+
+You can [create an issue](https://github.com/moead-framework/framework/issues/new) with all information that can 
+help us to understand your request. If you want to report a bug, share with us the following information:
+
+- version of python
+- version of the framework
+- operating system
+- the error raised by your code
+- all steps to reproduce the bug
+
+
+## You want to contribute to the framework code ?
+
+You can [create an issue](https://github.com/moead-framework/framework/issues/new) with information about the feature or the change you want to make. 
+You can then create a Pull Request with the following process :
 
 1. Update the README.md with details of changes to the interface, this includes new environment 
    variables and useful file locations.
-2. Update the documentation with a new Pull Request on this repository [https://github.com/moead-framework/framework/tree/master/docs](https://github.com/moead-framework/framework/tree/master/docs).
+2. Update the documentation available in the folder [https://github.com/moead-framework/framework/tree/master/docs](https://github.com/moead-framework/framework/tree/master/docs).
 3. Don't forget to add unit tests.
 4. Create a new branch and a Pull Request to merge your work on the **dev** branch of the project.
 
