### 2021-03-31 moead-framework/framework#13 @geoffreyp
I add more information in contributing.md and a section in readme : PR https://github.com/moead-framework/framework/pull/52
