### 2021-03-31 2b313c1
notice than set_seed is not safe in threaded environments
