+++ moead_framework/tool/result.py
@@ -46,7 +46,10 @@ def save_population_full(file_name, population):
 
 def set_seed(seed):
     """
-    Set the seed for numpy and random packages
+    Set the seed for numpy and random packages.
+    Warning, this approach is not safe in threaded environments.
+    By re-using a seed value, the same sequence should be reproducible from run to run as long as
+    multiple threads are not running (https://docs.python.org/3/library/random.html#notes-on-reproducibility)
 
     :param seed: integer
     :return:
