### 2021-03-31 moead-framework/framework#35 @geoffreyp
I understand the confusion but all the code is based on one function that creates an object OneDimensionSolution with the decision vector and the objective vector and it is not easy to refactor this part.

I propose other changes :

- rename `def generate_solution(self, array, evaluate=True)` by `def create_solution_object(self, decision_vector, evaluate=True)`
- rename the attribute `solution` by `decision_vector` as I said in the issue https://github.com/moead-framework/framework/issues/44
- Specify in the documentation that we consider the Solution object as an object with a decision vector and an objective vector

The word "generate" is used for me in this context to designate the creation of the object and not the generation of the decision vector, so I understand that the documentation must be more precise
