### 2021-03-31 68c505a
add tuto and example about results of algorithms
