+++ docs/tools.rst
@@ -1,0 +1,32 @@
+.. _components:
+
+Tools
+===========================================
+
+
+For multi-objective
+___________________________________________
+
+.. autosummary::
+   :nosignatures:
+   :toctree: moead_framework
+
+   moead_framework.tool.mop.get_non_dominated
+   moead_framework.tool.mop.is_duplicated
+   moead_framework.tool.mop.population_size_without_duplicate
+   moead_framework.tool.mop.compute_crowding_distance
+   moead_framework.tool.mop.get_non_dominated
+
+
+For results
+___________________________________________
+
+
+.. autosummary::
+   :nosignatures:
+   :toctree: moead_framework
+
+   moead_framework.tool.result.save_population
+   moead_framework.tool.result.save_population_full
+   moead_framework.tool.result.compute_hypervolume
+
