### 2021-03-31 moead-framework/framework#15 @geoffreyp
The documentation for tools is added in the PR https://github.com/moead-framework/framework/pull/55
The documentation for solution is available in the section Main Component: https://moead-framework.github.io/framework/html/main_components.html#solution
