### 2021-03-31 moead-framework/framework#33 @geoffreyp
I have a problem with this function, I don't need it and I don't know how to remove it from the documentation... I tried a lot of option in autodoc and autosummary without success
