### 2021-04-09 919c4b1
convert np.array to list in __repr__
