### 2021-04-14 moead-framework/framework#33 @sjvrijn
> I tried a lot of option in autodoc and autosummary without success

Did you also try the [negated form advanced options](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#directive-option-automodule-special-members)?

Meanwhile, I think the 'core issue' is that the aggregation functions don't specify their own `__init__` (since they don't have any state). It might disappear if you define your own at any point in the inheritance tree?

Although it also makes me wonder what the benefits are of creating stateless objects for this instead of simply having e.g. a file `tchebycheff.py` with functions `run()` and `is_better()`. After all, everything in Python is an object anyway, including imported modules, so you would still be able to use it as

```python
from moead_framework.aggregation import tchebycheff
tchebycheff.run(...)
tchebycheff.is_better(...)
```

So just wondering, did you consider this as an option?
