### 2021-04-15 moead-framework/framework#33 @geoffreyp
Thank you for the link to the documentation. I fixed the problem in the PR https://github.com/moead-framework/framework/pull/57. With autosummary, all class files are forced to add the init method even if it is not documented and without testing the flags no-undoc-members. So I edited the template directly with a custom template for aggregation functions.
