### 2021-04-23 f4fc38a
use custom template for Offspring generator
