+++ docs/moead_framework/moead_framework.aggregation.weighted_sum.WeightedSum.rst
@@ -6,7 +6,6 @@ moead\_framework.aggregation.weighted\_sum.WeightedSum
 .. autoclass:: WeightedSum
 
    
-   .. automethod:: __init__
 
    
    .. rubric:: Methods
