### 2021-05-05 31fd92b
Add Problem.evaluate() methods
