+++ moead_framework/problem/problem.py
@@ -1,4 +1,8 @@
 from abc import abstractmethod, ABC
+from typing import Union, Sequence
+
+from moead_framework.solution.solution import Solution
+from moead_framework.solution import OneDimensionSolution
 
 
 class Problem(ABC):
@@ -35,4 +39,13 @@ class Problem(ABC):
         :param array: {list<integer>} all decision variables of the Solution
         :param evaluate: {boolean} specify if the new solution is evaluated. The default value is True.
         :return: {:class:`~moead_framework.solution.one_dimension_solution.OneDimensionSolution`}
+        """
+
+    @abstractmethod
+    def evaluate(self, x: Union[Solution, Sequence]) -> OneDimensionSolution:
+        """
+        Evaluate the given solution for the current problem and store the outcome
+
+        :param x: A {Solution} containing all decision variables
+        :return: :class:`~moead_framework.solution.one_dimension_solution.OneDimensionSolution`
         """
