+++ moead_framework/problem/numerical/zdt.py
@@ -46,5 +46,10 @@ class Zdt1(Problem):
                 x.F.append(self.f(j, x.solution))
             else:
                 x.F.append(None)
+        return x
 
+    def evaluate(self, x):
+        if not isinstance(x, OneDimensionSolution):
+            x = OneDimensionSolution(x)
+        x.F = [self.f(j, x.solution) for j in range(self.number_of_objective)]
         return x
