+++ moead_framework/problem/combinatorial/rmnk.py
@@ -84,6 +84,12 @@ class Rmnk(Problem):
 
         return x
 
+    def evaluate(self, x):
+        if not isinstance(x, OneDimensionSolution):
+            x = OneDimensionSolution(np.array(x, dtype=int))
+        x.F = [self.f(j, x.solution) for j in range(self.number_of_objective)]
+        return x
+
     def load_links(self, file_content):
         """
         Load links from the instance file
