### 2021-05-05 8b28f3b
Refactor Problem.generate_solution -> Problem.evaluate()
