### 2021-05-05 5bc0289
Made Problem.evaluate() proper inherited method
Since the implementation in each Problem was almost identical
except for an additional np.array(x, dtype=int) call for the
RMNK and MUBQP problems, it was easily possible to define the
method under the base class 'Problem' instead.

This generic implementation includes the np.array(x) call, with
the dtype set on a per-class basis, with a default of 'float' to
make it as easy as possible for new problems as they only have to
override if the required dtype is different.
