+++ moead_framework/problem/problem.py
@@ -1,11 +1,14 @@
 from abc import abstractmethod, ABC
 from typing import Union, Sequence
 
+import numpy as np
 from moead_framework.solution.solution import Solution
 from moead_framework.solution import OneDimensionSolution
 
 
 class Problem(ABC):
+
+    dtype = float
 
     def __init__(self, objective_number=2):
         self.number_of_objective = objective_number
@@ -31,7 +34,6 @@ class Problem(ABC):
         :return: {:class:`~moead_framework.solution.one_dimension_solution.OneDimensionSolution`}
         """
 
-    @abstractmethod
     def evaluate(self, x: Union[Solution, Sequence]) -> OneDimensionSolution:
         """
         Evaluate the given solution for the current problem and store the outcome
@@ -39,3 +41,7 @@ class Problem(ABC):
         :param x: A {Solution} containing all decision variables
         :return: :class:`~moead_framework.solution.one_dimension_solution.OneDimensionSolution`
         """
+        if not isinstance(x, OneDimensionSolution):
+            x = OneDimensionSolution(np.array(x, dtype=self.dtype))
+        x.F = [self.f(j, x.solution) for j in range(self.number_of_objective)]
+        return x
