+++ moead_framework/problem/combinatorial/knapsack.py
@@ -106,9 +106,3 @@ class KnapsackProblem(Problem):
 
     def generate_random_solution(self, evaluate=True):
         return self.evaluate(x=np.random.randint(0, 2, self.number_of_objects).tolist()[:])
-
-    def evaluate(self, x):
-        if not isinstance(x, OneDimensionSolution):
-            x = OneDimensionSolution(x)
-        x.F = [self.f(j, x.solution) for j in range(self.number_of_objective)]
-        return x
