### 2021-05-05 moead-framework/framework#35 @sjvrijn
The name `create_solution_object(array)` suggests that the main function of that method is to perform the action
```
solution = OneDimensionalSolution(array)
```
Wrapping that in a separate function just makes it needlessly complicated.

In PR #58, I've implemented my proposed change of removing `generate_solution()` and introducing `evaluate()`, leaving out the change to `generate_random_solution`. This turned out to be easier than I too had expected. You can see in the usage in the algorithm and tests that end result of the refactor outside of the `Problem` classes is just a rename from `generate_solution` to `evaluate`🙂

I approve of your remaining suggestions, i.e., renaming `Solution.solution` and further clarifying the documentation.
