### 2021-05-08 7c88137
refactoring: rename OneDimensionSolution.solution and documentation
