+++ moead_framework/test/test_genetic_operator.py
@@ -38,7 +38,7 @@ class GeneticOperatorsTest(unittest.TestCase):
     def test_numerical_polynomial_mutation(self):
         """Test numerical polynomial mutation"""
         problem = Zdt1(20)
-        array_solution = problem.generate_random_solution().solution
+        array_solution = problem.generate_random_solution().decision_vector
         new_solution = PolynomialMutation(solutions=[array_solution]).run()
 
         self.assertEqual(new_solution, [0.13890427141075934, 0.95236226912859, 0.750939942185552, 0.21601593072335035,
@@ -51,9 +51,9 @@ class GeneticOperatorsTest(unittest.TestCase):
     def test_numerical_de_crossover(self):
         """Test numerical de crossover"""
         problem = Zdt1(5)
-        array_solution1 = problem.generate_random_solution().solution
-        array_solution2 = problem.generate_random_solution().solution
-        array_solution3 = problem.generate_random_solution().solution
+        array_solution1 = problem.generate_random_solution().decision_vector
+        array_solution2 = problem.generate_random_solution().decision_vector
+        array_solution3 = problem.generate_random_solution().decision_vector
 
         new_solution = DifferentialEvolutionCrossover(solutions=[array_solution1,
                                                                  array_solution2,
@@ -65,9 +65,9 @@ class GeneticOperatorsTest(unittest.TestCase):
     def test_moead_de_operator(self):
         """Test numerical de crossover"""
         problem = Zdt1(5)
-        array_solution1 = problem.generate_random_solution().solution
-        array_solution2 = problem.generate_random_solution().solution
-        array_solution3 = problem.generate_random_solution().solution
+        array_solution1 = problem.generate_random_solution().decision_vector
+        array_solution2 = problem.generate_random_solution().decision_vector
+        array_solution3 = problem.generate_random_solution().decision_vector
 
         new_solution = MoeadDeOperators(solutions=[array_solution1,
                                                    array_solution2,
