### 2021-05-09 moead-framework/framework#44 @chkoar
In Python we don't have to put each class in a separate package/module. Classes can co-exist in a single module/file. A `base.py` I think that is a common house for abstract/base classes or mixins. Well, naming things is a hard hobby. A solution is represented using a vector which is compose by values, right?
