+++ docs/moead_framework/moead_framework.problem.numerical.zdt.Zdt1.rst
@@ -14,9 +14,9 @@ moead\_framework.problem.numerical.zdt.Zdt1
    .. autosummary::
    
       ~Zdt1.__init__
+      ~Zdt1.evaluate
       ~Zdt1.f
       ~Zdt1.generate_random_solution
-      ~Zdt1.generate_solution
    
    
 
