### 2021-05-11 moead-framework/framework#15 @chkoar
Some links do not work. Check `Problem` and `AggregationFunction` [here](https://moead-framework.github.io/framework/html/moead_framework/moead_framework.algorithm.combinatorial.moead.Moead.html#moead_framework.algorithm.combinatorial.moead.Moead) for instance.
