### 2021-05-11 moead-framework/framework#15 @chkoar
1) There are classes, methods or functions that seem public and they do not contain a docstring. I list some below.
Please consider adding a descriptive docstring or make them "private" in case that is not intended for public use.

- [ ] `AggregationFunction`
- [ ] `Aggregation.run`
- [ ]  `MoeadDRA.run`
- [ ] `CrossoverAndMutation.run`
- [ ] `Crossover.run`
- [ ]  `BinaryMutation.run`
- [ ]  `DifferentialEvolutionCrossover.run`
- [ ]  `MoeadDeOperators.run`
- [ ] `PolynomialMutation.run`
- [ ] `OffspringGeneratorGeneric.run`
- [ ] `ParentSelector.run`
- [ ] `OneRandomAndCurrentParentSelector.select`
- [ ] `TwoRandomAndCurrentParentSelector.select`
- [ ] `TwoRandomParentSelector.select`
- [ ] `SpsStrategy`
- [ ] `SpsAllSubproblems.get_sub_problems`
- [ ] `MaxEvaluation.test`
- [ ] `KnapsackProblem.f`
- [ ] `KnapsackProblem.generate_random_solution`
- [ ] `Mubqp.generate_random_solution`
- [ ] `Rmnk.f`
- [ ] `Rmnk.generate_random_solution`
- [ ] `Problem`

2. It should be nice to include some execution examples in the class docstrings.
