### 2021-05-11 moead-framework/framework#60 @chkoar
__init__.py
Related: openjournals/joss-reviews#2974

Is there a reason for having a `__init__.py` file in the root of the repository?
