### 2021-05-11 moead-framework/framework#45 @chkoar
It is not fixed.
```
moead_framework/solution/one_dimension_solution.py:9: error: Need type annotation for 'decision_vector' (hint: "decision_vector: List[<type>] = ...")
moead_framework/solution/one_dimension_solution.py:10: error: Need type annotation for 'F' (hint: "F: List[<type>] = ...")
moead_framework/tool/result.py:67: error: Bracketed expression "[...]" is not valid as a type
moead_framework/tool/result.py:67: note: Did you mean "List[...]"?
moead_framework/tool/result.py:69: error: Incompatible types in assignment (expression has type "List[<nothing>]", variable has type "MultiList")
moead_framework/tool/result.py:71: error: Function "numpy.array" is not valid as a type
moead_framework/tool/result.py:71: note: Perhaps you need "Callable[...]" or a callback protocol?
moead_framework/tool/result.py:86: error: np.array? has no attribute "__iter__" (not iterable)
moead_framework/problem/combinatorial/rmnk.py:13: error: Incompatible types in assignment (expression has type "Type[int]", base class "Problem" defined the type as "Type[float]")
moead_framework/problem/combinatorial/mubqp.py:13: error: Incompatible types in assignment (expression has type "Type[int]", base class "Problem" defined the type as "Type[float]")
```

I propose to avoid the use of types if you are not familiar. Provide enough docstrings and it will be enough currently.
