### 2021-05-11 aedcca7
test if the parameter is an integer
