### 2021-05-11 moead-framework/framework#45 @geoffreyp
> I propose to avoid the use of types if you are not familiar. Provide enough docstrings and it will be enough currently.

what do I need to add in the docstring ? Can you give me an example here ?
