+++ moead_framework/algorithm/numerical/moead.py
@@ -7,6 +7,31 @@ from moead_framework.core.parent_selector.two_random_and_current_parent_selector
 class Moead(AbstractMoead):
     """
     Implementation of MOEA/D for numerical problems.
+
+
+    Example:
+
+    >>> from moead_framework.aggregation import Tchebycheff
+    >>> from moead_framework.algorithm.numerical import Moead
+    >>> from moead_framework.problem.numerical import Zdt1
+    >>>
+    >>> zdt = Zdt1(10)
+    >>>
+    >>> number_of_weight = 10
+    >>> number_of_weight_neighborhood = 2
+    >>> number_of_evaluations = 1000
+    >>> # The file is available here : https://github.com/moead-framework/data/blob/master/weights/SOBOL-2objs-10wei.ws
+    >>> # Others weights files are available here : https://github.com/moead-framework/data/tree/master/weights
+    >>> weight_file = "moead_framework/test/data/weights/SOBOL-" + str(zdt.number_of_objective) + "objs-" + str(number_of_weight) + "wei.ws"
+    >>>
+    >>> moead = Moead(problem=zdt,
+    >>>               max_evaluation=number_of_evaluations,
+    >>>               number_of_weight_neighborhood=number_of_weight_neighborhood,
+    >>>               weight_file=weight_file,
+    >>>               aggregation_function=Tchebycheff,
+    >>>               )
+    >>>
+    >>> population = moead.run()
     """
     def __init__(self, problem,
                  max_evaluation,
