+++ moead_framework/algorithm/combinatorial/moead_sps_random.py
@@ -5,6 +5,34 @@ from moead_framework.core.sps_strategy.sps_random_and_boundaries import SpsRando
 class MoeadSPSRandom(Moead):
     """
     Implementation of MOEA/D with the component :class:`~moead_framework.core.sps_strategy.sps_random_and_boundaries.SpsRandomAndBoundaries`.
+
+    Example:
+
+    >>> from moead_framework.aggregation import Tchebycheff
+    >>> from moead_framework.algorithm.combinatorial import MoeadSPSRandom
+    >>> from moead_framework.problem.combinatorial import Rmnk
+    >>>
+    >>> # The file is available here : https://github.com/moead-framework/data/blob/master/problem/RMNK/Instances/rmnk_0_2_100_1_0.dat
+    >>> # Others instances are available here : https://github.com/moead-framework/data/tree/master/problem/RMNK/Instances
+    >>> instance_file = "moead_framework/test/data/instances/rmnk_0_2_100_1_0.dat"
+    >>> rmnk = Rmnk(instance_file=instance_file)
+    >>>
+    >>> number_of_weight = 10
+    >>> number_of_weight_neighborhood = 2
+    >>> number_of_evaluations = 1000
+    >>> # The file is available here : https://github.com/moead-framework/data/blob/master/weights/SOBOL-2objs-10wei.ws
+    >>> # Others weights files are available here : https://github.com/moead-framework/data/tree/master/weights
+    >>> weight_file = "moead_framework/test/data/weights/SOBOL-" + str(rmnk.number_of_objective) + "objs-" + str(number_of_weight) + "wei.ws"
+    >>>
+    >>> moead = MoeadSPSRandom(problem=rmnk,
+    >>>               max_evaluation=number_of_evaluations,
+    >>>               number_of_subproblem_to_visit=1,
+    >>>               number_of_weight_neighborhood=number_of_weight_neighborhood,
+    >>>               weight_file=weight_file,
+    >>>               aggregation_function=Tchebycheff,
+    >>>               )
+    >>>
+    >>> population = moead.run()
     """
 
     def __init__(self, problem,
