### 2021-05-25 #2974 @chkoar
> @geoffreyp to summarize, there are still 5 issues to be reviewed by @chkoar and none for @sjvrijn.
Two of these issues are open questions that I answered a few months ago, 3 PR are created for the 3 other issues

I think that My review can be submitted regardless of these issues.
