### 2021-05-25 #2974 @chkoar
@melissawm till now, I have opened an issue in the software repo for each concern I had. Although there were not a direct match with the checklist. I have left some of the checks unchecked. Since some of them cannot be addressed by the author (e.g. General Checks), do you want me to paste here my report so we will be all aligned?
