### 2021-05-27 #2974 @sjvrijn
Hi @melissawm, apologies for the recent lack of response & progress. I hope to complete my review later today by raising all remaining issues I see for me to complete the checklist.

One question regarding the software paper though: Besides just multi-objective optimization software, other frameworks also exist for single-objective optimization such as [DEAP](https://deap.readthedocs.io/), and a framework for CMA-ES, similarly modular to this project, I worked on ([ModEA](https://github.com/sjvrijn/ModEA)). Are these structurally relevant enough to include in the paper, or is it enough for the paper to focus on multi-objective optimization only?
