### 2021-05-27 moead-framework/framework#62 @sjvrijn
[joss-reviews#2974] Improve documentation/README for support
Related: openjournals/joss-reviews#2974

This issues concerns two points:

- The main page of the documentation instructs readers to raise an issue, while the CONTRIBUTING.md suggests to open a discussion.
- The README does not give any direction when searching for terms like 'support' or 'question'.

Please make the instruction on where to ask questions consistent add a (sub)section 'Support' in the README directing users there if they have any questions.
