### 2021-05-27 moead-framework/framework#64 @sjvrijn
[joss-reviews#2974] Checkpoints used inconsistently
Related: openjournals/joss-reviews#2974 and #15

See e.g. the difference between
https://github.com/moead-framework/framework/blob/659cb0404b356f80a30da83e63ff443104e4f2f2/moead_framework/algorithm/abstract_moead.py#L103-L104
and
https://github.com/moead-framework/framework/blob/659cb0404b356f80a30da83e63ff443104e4f2f2/moead_framework/algorithm/combinatorial/moead_dra.py#L69-L70
