### 2021-05-27 moead-framework/framework#64 @sjvrijn
Instead of the externally referenced version given in the [tutorial section](https://moead-framework.github.io/framework/html/tuto.html?highlight=checkpoint#save-data-with-the-framework), it is probably better to have a checkpoint function always accept the MOEAD algorithm object as only argument. This would allow for a generic checkpoint function to be written without it being tied to a single MOEAD instance. I.e.
```python
def checkpoint(moead_algorithm: AbstractMoead):
    ....  # perform checkpoint actions here
```
with the call then becoming
```python
if checkpoint:
    checkpoint(self)
```

A simple checkpoint function could then for example be added to the `tools` submodule as an example.
