+++ docs/contributing.rst
@@ -2,7 +2,7 @@
 Contributing
 ===========================================
 
-When contributing to this project, please first discuss the change you wish to make via issue,
+When contributing to this project, please first discuss the change you wish to make via GitHub discussions,
 email, or any other method with the owners of this repository before making a change. 
 
 Please note we have a code of conduct, please follow it in all your interactions with the project.
@@ -14,10 +14,22 @@ ________________________________
    build.
 2. Update the README.md with details of changes to the interface, this includes new environment 
    variables, exposed ports, useful file locations and container parameters.
-3. Update the documentation with a new Pull Request on the repository https://github.com/moead-framework/documentation.
+3. Update the documentation in the same Pull Request
 4. Don't forget to add unit tests.
 5. Create a new branch and a Pull Request to merge your work on the **dev** branch of the project.
 
+
+You want to share a bug ?
+________________________________
+
+You can create an issue with all information that can
+help us to understand your request. Share with us the following information:
+
+- version of python
+- version of the framework
+- operating system
+- the error raised by your code
+- all steps to reproduce the bug
 
 Code of Conduct
 ________________________________
