### 2021-05-28 4717ec3
update checkpoint documentation
