+++ docs/main_components.rst
@@ -19,6 +19,12 @@ __________________________________________
 
 These algorithms are implemented on the basis of the original papers (:cite:`moead`, :cite:`moead_de`, :cite:`moead_dra`, :cite:`gpruvost_evocop2020`).
 
+.. autosummary::
+   :nosignatures:
+   :toctree: moead_framework
+
+   moead_framework.algorithm.abstract_moead.AbstractMoead
+
 For combinatorial problems
 ---------------------------------
 
