+++ docs/tuto.rst
@@ -123,10 +123,10 @@ If you want to save all non-dominated solutions (attribute :code:`self.ep` in th
               aggregation_function=Tchebycheff
               )
 
-    def checkpt():
-        if moead.current_eval % 10 ==0 :
-            filename = "non_dominated_solutions-eval" + str(moead.current_eval) + ".txt"
-            save_population(file_name=filename, population=moead.ep)
+    def checkpoint(moead_algorithm: AbstractMoead)::
+        if moead_algorithm.current_eval % 10 ==0 :
+            filename = "non_dominated_solutions-eval" + str(moead_algorithm.current_eval) + ".txt"
+            save_population(file_name=filename, population=moead_algorithm.ep)
 
     moead.run(checkpoint=checkpt)
 
