### 2021-05-28 70ea112
fix knapsack documentation
