### 2021-05-28 moead-framework/framework#64 @geoffreyp
The documentation and the code are updated in the PR #55 (commit 4717ec3)
