### 2021-05-28 7a9c614
add more informations about compatibility between each components
