+++ moead_framework/core/selector/delta_selector.py
@@ -5,9 +5,24 @@ from .abstract_selector import MatingPoolSelector
 
 class DeltaSelector(MatingPoolSelector):
     """
-    According to the probability delta (attribute in the algorithm), the selector will select the neighborhood or the whole population.
-    delta is the probability that parent solutions are selected from the neighborhood.
+    According to the probability delta (required attribute in the algorithm), the selector will select the neighborhood or the whole population.
+    The attribute 'delta' represents the probability that parent solutions are selected from the neighborhood.
     """
+
+    def __init__(self, algorithm_instance):
+        """
+        Constructor of the mating pool selector
+
+        :param algorithm_instance: {:class:`~moead_framework.algorithm.abstract_moead.py.AbstractMoead`} instance of the algorithm
+        """
+        super().__init__(algorithm_instance)
+
+        try:
+            if not hasattr(self.algorithm, 'delta'):
+                raise
+        except Exception:
+            msg = "The algorithm does not seem to be compatible with the component DeltaSelector. The parameter delta is required in the algorithm by the component DeltaSelector."
+            raise ValueError(msg)
 
     def select(self, sub_problem):
         """
