+++ moead_framework/core/genetic_operator/abstract_operator.py
@@ -28,5 +28,6 @@ class GeneticOperator(ABC):
         :param n: {integer} number of solution required by the operator
         :return:
         """
-        if len(self.solutions) < n:
-            raise ValueError("The genetic operator needs more than "+str(n)+" solutions")
+        if len(self.solutions) != n:
+            msg = "The algorithm does not seem to be compatible with the current genetic operator. The current genetic oeperator needs {n} solutions as parents and it received {badn} solution(s)."
+            raise ValueError(msg.format(n=n, badn=len(self.solutions)))
