### 2021-05-28 moead-framework/framework#63 @geoffreyp
All components are not compatible with each algorithm because they have some requirements.
For example :
 - `moead_framework.core.sps_strategy.sps_dra.SpsDra` needs the attribute pi
 - `SpsRandomAndBoundaries` needs the attribute number_of_subproblem
 -  The multi-point Crossover needs 2 solutions as parents so we have to use a mating selector that give us 2 parents
 -  ...

I added more information in the documentation and I improved errors to say how the component ... is not compatible with the current algorithm. PR #66
