### 2021-05-28 632be94
fix name of checkpoint function in the tutorial page
