### 2021-05-31 moead-framework/framework#18 @chkoar
`moead_framework/algorithm/abstract_moead.py [106:112]`:
```python
  self.update_current_sub_problem(sub_problem=i)
  self.mating_pool = self.mating_pool_selection(sub_problem=i)[:]
  y = self.generate_offspring(population=self.mating_pool)
  y = self.repair(solution=y)
  self.update_z(solution=y)
  self.update_solutions(solution=y, aggregation_function=self.aggregation_function, sub_problem=i)
  self.current_eval += 1
```

`moead_framework/algorithm/combinatorial/moead_dra.py [72:78]`:
```python
  self.update_current_sub_problem(sub_problem=i)
  self.mating_pool = self.mating_pool_selection(sub_problem=i)[:]
  y = self.generate_offspring(population=self.mating_pool)
  y = self.repair(solution=y)
  self.update_z(solution=y)
  self.update_solutions(solution=y, aggregation_function=self.aggregation_function, sub_problem=i)
  self.current_eval += 1
```
