### 2021-06-01 moead-framework/framework#63 @sjvrijn
Thanks for elaborating that not all components are expected to work together. Here are my suggestions to address this issue:

- **Fix numerical and combinatorial Moead**
Lines [68 in combinatorial](https://github.com/moead-framework/framework/blob/master/moead_framework/algorithm/combinatorial/moead.py#L68) and [65 in numerical](https://github.com/moead-framework/framework/blob/master/moead_framework/algorithm/numerical/moead.py#L65) Moead still need to initialize the selector with `parent_selector(algorithm=self)`

   The if-else block may even be simplified to the example below for the combinatorial, and equivalently for the numerical case:
```python
if parent_selector is None:
    parent_selector = TwoRandomParentSelector
self.parent_selector = parent_selector(algorithm=self)
```

- **Documentation:**
  In the global documentation, I would add a clear generic note stating something like
  > **Note**: Not all available components are compatible with each other.
  > Refer to each component's documentation for more details on required attributes.

  at the _Tutorials / Implement own algorithm_ and _Library Reference / Other components_ pages

- **Exceptions:**
  I'll post my explicit suggestions in the PR you already opened
