### 2021-06-02 1ed210f
remove duplication code in algorithm and sps strategies
