### 2021-06-02 233c694
fix attribute error messages
