### 2021-06-02 de0285d
update documentation and algorithm (selector component)
