### 2021-06-02 moead-framework/framework#63 @geoffreyp
Thank you for your suggestions !
The PR is updated with the commits 233c694 and de0285d
