### 2021-06-03 moead-framework/framework#64 @sjvrijn
This issue can be closed when PR #55 is merged
