### 2021-06-28 #2974 @geoffreyp
Hello @melissawm @sjvrijn @chkoar,
I have completed the requested changes. Do you think that the review can be finished before August? It's just to know if I can include this work in my thesis before sending it :)
