### 2021-06-28 #2974 @sjvrijn
Hi @geoffreyp,
I can check 'functionality documentation' once moead-framework/framework#55 has been approved by @chkoar and merged.

For 'state of the field' I'm waiting for a resolution my question about other frameworks from May 27th: either adding a paragraph on them in your paper, or a confirmation from @melissawm that they are out-of-scope for this submission.
