### 2021-06-28 #2974 @chkoar
@melissawm hey! As I [said earlier](https://github.com/openjournals/joss-reviews/issues/2974#issuecomment-869566182) I will open some issues in the software repo, so the review will move forward. However, some checks cannot be addressed from @geoffreyp. For instance the General Checks that I haven't ticked yet. Could you please see my comments and validate that they can be ticked without any concern?

# General checks

***Contribution and authorship: Has the submitting author (@geoffreyp) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?***

The submitting author [seems](https://github.com/openjournals/joss-reviews/issues/2895#issuecomment-741149229) to be the solely contributor to this package. However, the paper states two other authors. The problem here is that I cannot see and validate their contribution to this package. I can see their names in a cited papers, though.

**Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines**

JOSS requires that the published software should represent substantial scholarly effort. Thus, it lists some [factors](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) that should be considered.

At the [date](https://github.com/openjournals/joss-reviews/issues/2895#issuecomment-741149229) **of the submission** the submitted project had 60 commits, 1 author and almost 1.5k lines of Python code (549 blank and 236 comments)

Although, the submitted software seems that it have been used in two publications of the submitted author.
