### 2021-06-29 31f00dc
Revert "add examples of modular frameworks"
This reverts commit 18bfe45a7e5020660c490dc090fd6ee11adc886f.
