+++ paper.bib
@@ -118,21 +118,3 @@ keywords = {open source, multi-objective metaheu- ristics, optimization framewor
 location = {Madrid, Spain},
 series = {GECCO Companion '15}
 }
-
-@INPROCEEDINGS{vanrijn2016,
-    author={S. {van Rijn} and H. {Wang} and M. {van Leeuwen} and T. {Bäck}},
-    booktitle={2016 IEEE Symposium Series on Computational Intelligence (SSCI)},
-    title={Evolving the structure of Evolution Strategies},
-    year={2016},
-    doi={10.1109/SSCI.2016.7850138},
-}
-
-@article{DEAP_JMLR2012,
-    author    = " F\'elix-Antoine Fortin and Fran\c{c}ois-Michel {De Rainville} and Marc-Andr\'e Gardner and Marc Parizeau and Christian Gagn\'e ",
-    title     = { {DEAP}: Evolutionary Algorithms Made Easy },
-    pages     = { 2171--2175 },
-    volume    = { 13 },
-    month     = { jul },
-    year      = { 2012 },
-    journal   = { Journal of Machine Learning Research }
-}
