### 2021-06-29 18bfe45
add examples of modular frameworks
