### 2021-06-29 #2974 @geoffreyp
Hello @sjvrijn, you can see my answer in the PR https://github.com/moead-framework/framework/pull/68.

> Hey @geoffreyp, I will open three issues so you can address them in order to move forward the review.

Hello @chkoar, thank you, so I will wait for the new issues !

Concerning the **Contribution and authorship**, Bilel and Arnaud are my supervisors, and they contributed to the framework in its design and in the writing of the paper, even if it is true that they do not appear in the contributors on GitHub
