### 2021-07-05 moead-framework/framework#69 @chkoar
[joss-reviews#2974] Functionality
Related: openjournals/joss-reviews#2974

**Functionality: Have the functional claims of the software been confirmed?**
The authors in the paper state that:

> With the moead-framework package, we aim at **(1)** bringing the modularity of the MOEADr package by using the flexibility of Python in order to allow the user to update the behavior of MOEA/D components in their research works, and **(2)** to propose new variants without being limited by the software.

The authors should provide detailed code examples of (1) and (2). Unless I am missing something.
