### 2021-07-05 moead-framework/framework#70 @chkoar
[joss-reviews#2974] Software paper
Related: openjournals/joss-reviews#2974

**State of the field: Do the authors describe how this software compares to other commonly-used packages?**

In the submitted paper the authors write that "These software offer many state-of-art algorithms, visualization tools or parallelization abstraction but they do not enable to test and analyze in detail the behavior of components implemented in each algorithm"

I cannot understand this statement. @geoffreyp could you elaborate more on this?

cc @sjvrijn I think that is related to https://github.com/openjournals/joss-reviews/issues/2974#issuecomment-869755184
