### 2021-07-05 moead-framework/framework#71 @chkoar
[joss-reviews#2974] Documentation
Related: openjournals/joss-reviews#2974

**(1) Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

There is [an example](https://moead-framework.github.io/framework/html/examples.html) script. Although, I would expect to see several examples that will demonstrate the major components of the software.

~~**(2) Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?**~~

~~I think that a user guide is missing. User tutorials have been added but I think that they need polishing.~~

**(3) Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?**

The code coverage is great 95%. Although, as I can see the tests cover only the happy path. For instance if you pass not expected types the user should guess what is happening.

~~**(4) Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**~~

~~I cannot see (2) and (3)~~
