### 2021-07-05 #2974 @chkoar
Hey @geoffreyp,

I have opened moead-framework/framework#69, moead-framework/framework#69 and moead-framework/framework#70. These are my objections regarding my unticked items in my report.

@melissawm I suppose that @geoffreyp's [answer](https://github.com/openjournals/joss-reviews/issues/2974#issuecomment-870455114) is adequate in order to tick the [General Checks](https://github.com/openjournals/joss-reviews/issues/2974#issuecomment-869782530), right?

cc @sjvrijn
