### 2021-07-06 #2974 @melissawm
> _**Contribution and authorship: Has the submitting author (@geoffreyp) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?**_
>
> The submitting author [seems](https://github.com/openjournals/joss-reviews/issues/2895#issuecomment-741149229) to be the solely contributor to this package. However, the paper states two other authors. The problem here is that I cannot see and validate their contribution to this package. I can see their names in the cited papers, though.

From the JOSS Documentation [here](https://joss.readthedocs.io/en/latest/submitting.html#authorship) and [here](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship)

*Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, **but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed.** In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.*

It is reasonable to expect that the co-authors are then aware and agree to be listed here.

As far as Substantial Scholarly Effort, I don't see an objection. (although of course Reviewers are free to disagree :)
