### 2021-07-09 moead-framework/framework#69 @geoffreyp
With the framework, you can directly extend the algorithms to change the interaction between each component by rewriting the run() method of the algorithm, or use the modularity of the framework to use your own components by using them as parameters of the algorithm's constructor.  It is explained in the following paragraph, I think the pdf is not updated but this paragraph was already updated with @sjvrijn in https://github.com/moead-framework/framework/pull/30 :

> With the *moead-framework* package, we aim to provide the modularity of the MOEADr package by using the flexibility of Python. Indeed, we want to allow the user to update the behavior of MOEA/D components in their research works without being limited by the software. The package is focused on a modular architecture for easily adding, updating or testing the components of MOEA/D and for customizing how components interact with each other. Indeed, in contrast with other existing implementations, *moead-framework* does not limit the users with a limited number of components available as parameters (8 components are available in MOEADr). Users can easily restructure the 10 existing components of the *moead-framework* and include new ones to easily add new features without altering existing components. Components are not only customizable with parameters as with MOEADr, but in fact they can be added with the inheritance mechanism on the main run() method of each algorithm.
