### 2021-07-09 moead-framework/framework#70 @geoffreyp
The paper was already updated with @sjvrijn (but not the pdf I think) :

> The first version of MOEA/D and its most famous variants [@moead_de; @moead_dra] are implemented in recent multi-objective optimization software such as pymoo [@pymoo], pygmo [@pygmo] and jMetal [@jmetal]. These implementations offer many state-of-the-art algorithms, visualization tools or parallelization abstraction, but they do not enable detailed testing and analysis of the various algorithm's components' behavior.
