### 2021-07-09 moead-framework/framework#71 @geoffreyp
> There is [an example](https://moead-framework.github.io/framework/html/examples.html) script. Although, I would expect to see several examples that will demonstrate the major components of the software.

I made the choice to propose only one complete example with the initialization of a problem and an algorithm, the execution of an algorithm, the saving of the results and the extraction of information. Would you prefer that I decompose it into several examples?

> I think that a user guide is missing. User tutorials have been added but I think that they need polishing.

What kind of information would you like to see in a user guide?

> The code coverage is great 95%. Although, as I can see the tests cover only the happy path. For instance if you pass not expected types the user should guess what is happening.

I don't understand the request. Do you have an example of a test to add? The information about the expected types are in the documentation for each function.

> **Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**
>
> I cannot see (2) and (3)

This information is available in the Readme and in the documentation :
