### 2021-07-19 #2974 @chkoar
> As far as Substantial Scholarly Effort, I don't see an objection. (although of course Reviewers are free to disagree :)

OK. Checked.
