### 2021-07-19 moead-framework/framework#70 @chkoar
> These implementations offer many state-of-the-art algorithms, visualization tools or parallelization abstraction, but they do not enable **detailed testing** and **analysis of** the various algorithm's **components' behavior.**

Since this is the comparative advantage of MOEA/D I think that you should show it in the paper.
