### 2021-07-19 moead-framework/framework#69 @chkoar
> With the moead-framework package, we aim at (1) bringing the modularity of the MOEADr package by using the flexibility of Python in order to allow the user to update the behavior of MOEA/D components in their research works, and (2) to propose new variants without being limited by the software.

My point is that these lines are written in the paper but I cannot see code examples that demonstrate the above statement. I would like to see detailed code examples for both cases.
