### 2021-07-19 moead-framework/framework#71 @chkoar
**(1)**
I would like to see specific examples that will show the value of the library. Take an example what we are discussing [here](https://github.com/moead-framework/framework/issues/69#issuecomment-882625819). Create a new variant without being limited by the software and show the performance benefit using the new variant. [Sphinx-Gallery](https://sphinx-gallery.github.io/stable/index.html) might help you to show this.

**(2)**
DEAP makes a walk though in their [basic tutorials](https://deap.readthedocs.io/en/master/tutorials/basic/part1.html) showing how every components can be used in order to solve a real optimization problem.

**(3)**
> I don't understand the request. Do you have an example of a test to add? The information about the expected types are in the documentation for each function.

I was trying to say that the happy path is now always enough.
```python
In [8]: KnapsackProblem(1,10)

OSError: [Errno 6] Device not configured
```

**(4)**
Sorry I missed this. I was looking on the documentation pages [here](https://moead-framework.github.io/framework/html/contributing.html). Could you please also add the information there? There are no links.
