### 2021-07-19 moead-framework/framework#16 @chkoar
My point is that requiring external files (even for known problems) makes the things difficult. Take a look [here](https://github.com/jMetal/jMetalPy/blob/master/jmetal/problem/singleobjective/knapsack.py). My proposal is to use instance files directly as a last resort. 1) provide the data directly (I know cumbersome), 2) Using a factory and 3)instance files.
