### 2021-07-19 moead-framework/framework#16 @chkoar
> makes the things difficult.

Even for the example in the readme a user is instructed to go somewhere and download a file.
