### 2021-07-19 moead-framework/framework#45 @chkoar
> Thank you for the example, but what do I need to add in my documentation ?

As long as you keep your documentation in good coverage probably you will be ok without annotations.
Although, having annotations is a good thing for several reasons.

The following problems remain.

```
moead_framework/solution/one_dimension_solution.py:9: error: Need type annotation for 'decision_vector' (hint: "decision_vector: List[<type>] = ...")
moead_framework/solution/one_dimension_solution.py:10: error: Need type annotation for 'F' (hint: "F: List[<type>] = ...")
moead_framework/tool/result.py:82: error: Bracketed expression "[...]" is not valid as a type
moead_framework/tool/result.py:82: note: Did you mean "List[...]"?
moead_framework/tool/result.py:84: error: Incompatible types in assignment (expression has type "List[<nothing>]", variable has type "MultiList")
moead_framework/tool/result.py:86: error: Function "numpy.array" is not valid as a type
moead_framework/tool/result.py:86: note: Perhaps you need "Callable[...]" or a callback protocol?
moead_framework/tool/result.py:101: error: np.array? has no attribute "__iter__" (not iterable)
moead_framework/problem/problem.py:44: error: Argument 1 to "array" has incompatible type "Union[Solution, Sequence[Any]]"; expected "Union[List[Any], List[List[Any]], List[List[List[Any]]], List[List[List[List[Any]]]], Iterable[ndarray[Any]], ndarray[Any]]"
moead_framework/problem/combinatorial/rmnk.py:13: error: Incompatible types in assignment (expression has type "Type[int]", base class "Problem" defined the type as "Type[float]")
moead_framework/problem/combinatorial/mubqp.py:13: error: Incompatible types in assignment (expression has type "Type[int]", base class "Problem" defined the type as "Type[float]")
Found 9 errors in 5 files (checked 71 source files)
```
