+++ docs/moead_framework/moead_framework.tool.mop.is_duplicated.rst
@@ -1,0 +1,6 @@
+moead\_framework.tool.mop.is\_duplicated
+========================================
+
+.. currentmodule:: moead_framework.tool.mop
+
+.. autofunction:: is_duplicated
