+++ docs/moead_framework/moead_framework.aggregation.functions.AggregationFunction.rst
@@ -1,0 +1,23 @@
+moead\_framework.aggregation.functions.AggregationFunction
+==========================================================
+
+.. currentmodule:: moead_framework.aggregation.functions
+
+.. autoclass:: AggregationFunction
+
+   
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~AggregationFunction.__init__
+      ~AggregationFunction.is_better
+      ~AggregationFunction.run
+   
+   
+
+   
+   
+   
