+++ docs/moead_framework/moead_framework.core.selector.abstract_selector.MatingPoolSelector.rst
@@ -1,0 +1,23 @@
+moead\_framework.core.selector.abstract\_selector.MatingPoolSelector
+====================================================================
+
+.. currentmodule:: moead_framework.core.selector.abstract_selector
+
+.. autoclass:: MatingPoolSelector
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~MatingPoolSelector.__init__
+      ~MatingPoolSelector.select
+   
+   
+
+   
+   
+   
