+++ docs/moead_framework/moead_framework.tool.mop.get_non_dominated.rst
@@ -1,0 +1,6 @@
+moead\_framework.tool.mop.get\_non\_dominated
+=============================================
+
+.. currentmodule:: moead_framework.tool.mop
+
+.. autofunction:: get_non_dominated
