+++ docs/moead_framework/moead_framework.tool.result.save_population.rst
@@ -1,0 +1,6 @@
+moead\_framework.tool.result.save\_population
+=============================================
+
+.. currentmodule:: moead_framework.tool.result
+
+.. autofunction:: save_population
