+++ docs/moead_framework/moead_framework.core.offspring_generator.abstract_mating.OffspringGenerator.rst
@@ -1,0 +1,22 @@
+moead\_framework.core.offspring\_generator.abstract\_mating.OffspringGenerator
+==============================================================================
+
+.. currentmodule:: moead_framework.core.offspring_generator.abstract_mating
+
+.. autoclass:: OffspringGenerator
+
+   
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~OffspringGenerator.__init__
+      ~OffspringGenerator.run
+   
+   
+
+   
+   
+   
