+++ docs/moead_framework/moead_framework.algorithm.abstract_moead.AbstractMoead.rst
@@ -1,0 +1,34 @@
+moead\_framework.algorithm.abstract\_moead.AbstractMoead
+========================================================
+
+.. currentmodule:: moead_framework.algorithm.abstract_moead
+
+.. autoclass:: AbstractMoead
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~AbstractMoead.__init__
+      ~AbstractMoead.generate_closest_weight_vectors
+      ~AbstractMoead.generate_offspring
+      ~AbstractMoead.get_sub_problems_to_visit
+      ~AbstractMoead.init_z
+      ~AbstractMoead.initial_population
+      ~AbstractMoead.mating_pool_selection
+      ~AbstractMoead.optimize_sub_problem
+      ~AbstractMoead.repair
+      ~AbstractMoead.run
+      ~AbstractMoead.update_current_sub_problem
+      ~AbstractMoead.update_solutions
+      ~AbstractMoead.update_z
+   
+   
+
+   
+   
+   
