+++ docs/moead_framework/moead_framework.core.genetic_operator.abstract_operator.GeneticOperator.rst
@@ -1,0 +1,24 @@
+moead\_framework.core.genetic\_operator.abstract\_operator.GeneticOperator
+==========================================================================
+
+.. currentmodule:: moead_framework.core.genetic_operator.abstract_operator
+
+.. autoclass:: GeneticOperator
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~GeneticOperator.__init__
+      ~GeneticOperator.number_of_solution_is_correct
+      ~GeneticOperator.run
+   
+   
+
+   
+   
+   
