+++ docs/moead_framework/moead_framework.tool.result.compute_hypervolume.rst
@@ -1,0 +1,6 @@
+moead\_framework.tool.result.compute\_hypervolume
+=================================================
+
+.. currentmodule:: moead_framework.tool.result
+
+.. autofunction:: compute_hypervolume
