### 2021-07-21 moead-framework/framework#69 @geoffreyp
The two new examples are available inr the PR https://github.com/moead-framework/framework/pull/72
