### 2021-07-28 moead-framework/framework#71 @geoffreyp
New examples are available in the PR : https://github.com/moead-framework/framework/pull/72
The user guide is available in the PR https://github.com/moead-framework/framework/pull/73
The page contributing was updated with information from README.md
