### 2021-07-28 moead-framework/framework#70 @geoffreyp
I suggest adding a sentence:

> These implementations offer many state-of-the-art algorithms, visualization tools or parallelization abstraction, but they do not enable detailed testing and analysis of the various algorithm's components' behavior.
> In these libraries, the algorithms are seen as components that do not have sub-components, which makes their parameterization and analysis more difficult, we cannot easily modify their behavior.
