### 2021-07-28 moead-framework/framework#70 @chkoar
> In these libraries, the algorithms are seen as components that do not have sub-components, which makes their parameterization and analysis more difficult, we cannot easily modify their behavior.

I think that this is probably related to #69. Apart from the new text, can you please provide a concrete example of an idea that cannot be implemented with a commonly-used package but can easily be implemented with MOEA/D?
