### 2021-08-02 moead-framework/framework#71 @chkoar
> New examples are available in the PR : #72
The user guide is available in the PR #73
The page contributing was updated with information from README.md

Can you please render the updated documentation?
