### 2021-08-02 moead-framework/framework#69 @chkoar
Can you please render the updated documentation?

In each example script please provide a module docstring so the reader will be able to understand the purpose of each example.
