### 2021-08-16 6293f29
add a module docstring to each examples
