### 2021-08-16 moead-framework/framework#71 @geoffreyp
To render the documentation, I have to merge the PR in Master.
The documentation can be generated locally if you want check changes. The documentation is generated with sphinx 2.4.4 (see the section 'Requirements for developers' of the README).

You can generate the documentation with the following commands :

    cd docs/

    make html
