### 2021-08-16 moead-framework/framework#70 @geoffreyp
With pygmo and Jmetal for example, users are very free to implement new algorithms because they are not restricted by subcomponents, but this also makes it more difficult to implement new algorithms based on MOEA/D. With these frameworks, it is necessary to reimplement a complete algorithm (with code duplication) to implement variants of MOEA/D whereas with our framework, it is only sufficient to implement the component or the function that we want to modify or study. Each of the examples proposed in the documentation cannot therefore be implemented more easily than with our framework.
