+++ moead_framework/problem/combinatorial/rmnk.py
@@ -1,6 +1,5 @@
 import numpy as np
 from moead_framework.problem.problem import Problem
-from moead_framework.solution.one_dimension_solution import OneDimensionSolution
 
 
 class Rmnk(Problem):
@@ -30,12 +29,47 @@ class Rmnk(Problem):
 
     dtype = int
 
-    def __init__(self, instance_file):
-        """
-        Constructor of the problem
+    def __init__(self, instance_file=None, rho=None, m=None, n=None, k=None, links=None, tables=None):
+        f"""
+        Constructor of the problem. 
+        You can initialize the problem directly by using an instance file or by setting parameters : rho, m, n, k, links and tables. 
+        http://mocobench.sourceforge.net/index.php?n=Problem.RMNK#Format
 
-        :param instance_file: {string} text file generated by the rmnk generator : http://mocobench.sourceforge.net/index.php?n=Problem.RMNK#Code
+        :param instance_file: {str} text file generated by the rmnk generator : http://mocobench.sourceforge.net/index.php?n=Problem.RMNK#Code
+        :param rho: {float} the objective correlation coefficient
+        :param m: {int} the number of objective functions
+        :param n: {int} the length of solutions
+        :param k: {list} the number of epistatic links (non-linearity)
+        :param links: {list} describe the epistatic structure between variables
+        :param tables: {list} describe the fitness contribution
         """
+        self.rho = None
+        self.m = None
+        self.n = None
+        self.k = None
+        self.links = None
+        self.tables = None
+        self.instance_file = None
+
+        if instance_file is not None:
+            self.init_with_instance_file(instance_file=instance_file)
+        elif (rho is not None) & (m is not None) & (n is not None) & (k is not None) & (links is not None) & (
+                tables is not None):
+            self.init_with_data(rho=rho, m=m, n=n, k=k, links=links, tables=tables)
+        else:
+            msg = "The constructor needs either the instance_file parameter or the following parameters : " \
+                  "rho, m, n, k, links and tables."
+            raise ValueError(msg)
+
+        super().__init__(objective_number=self.m)
+
+    def init_with_instance_file(self, instance_file):
+
+        if isinstance(instance_file, str):
+            self.instance_file = instance_file
+        else:
+            raise TypeError("The instance_file parameter must be a string.")
+
         file = open(instance_file, 'r')
         file_content = list(map(str.strip, file.readlines()))
         file_content = file_content[3:]
@@ -52,14 +86,33 @@ class Rmnk(Problem):
 
         self.tables = np.zeros((self.m, self.n, 1 << (self.k + 1)))
 
-
         file_content = file_content[2:]
         line = self.load_links(file_content)
 
-        file_content = file_content[line+1:]
+        file_content = file_content[line + 1:]
         self.load_tables(file_content)
 
         file.close()
+
+    def init_with_data(self, rho, m, n, k, links, tables):
+
+        if isinstance(m, int) & isinstance(n, int) & isinstance(k, int):
+            self.m = m
+            self.n = n
+            self.k = k
+        else:
+            raise TypeError("The parameters m, n and k must be positive integers.")
+
+        if isinstance(rho, (int, float)):
+            self.rho = rho
+        else:
+            raise TypeError("The parameter rho must be a float.")
+
+        if isinstance(links, list) & isinstance(tables, list):
+            self.links = links
+            self.tables = tables
+        else:
+            raise TypeError("The parameters links and tables must be list.")
 
     def f(self, function_id, decision_vector):
         """
