### 2021-09-15 b2d93e7
fix annotation error about dtype
