### 2021-09-15 moead-framework/framework#45 @geoffreyp
It seems to be solved in https://github.com/moead-framework/framework/pull/75.
I only have errors on my side with numpy stubs

```
(framework-dev) ➜  framework git:(issue-45-error-annotation) ✗ mypy moead_framework
moead_framework/tool/mop.py:2: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/core/genetic_operator/combinatorial/crossover.py:2: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/problem/problem.py:4: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/core/genetic_operator/numerical/moead_de_operators.py:2: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/algorithm/abstract_moead.py:3: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/problem/numerical/zdt.py:1: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/problem/combinatorial/rmnk.py:1: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/problem/combinatorial/mubqp.py:1: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/problem/combinatorial/knapsack.py:1: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/tool/result.py:3: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/algorithm/combinatorial/moead_dra.py:1: error: Skipping analyzing 'numpy': found module but no type hints or library stubs
moead_framework/algorithm/combinatorial/moead_dra.py:1: note: See https://mypy.readthedocs.io/en/latest/running_mypy.html#missing-imports
Found 11 errors in 11 files (checked 71 source files)

```
