### 2021-09-25 moead-framework/framework#45 @chkoar
> It seems to be solved in #75.

Yes, it does.
