### 2021-09-25 moead-framework/framework#71 @chkoar
This issue raise concerns regarding 4 checks. Can you please point me where (PR/branch) each concern is addressed? Thanks.
