### 2021-09-25 moead-framework/framework#69 @chkoar
>  to propose new variants without being limited by the software

I would remove that sentence or I would replace it with something like this: `to propose new variants without being limited by the package itself.`
