### 2021-09-25 moead-framework/framework#70 @chkoar
> With pygmo and Jmetal for example, users are very free to implement new algorithms because they are not restricted by subcomponents, but this also makes it more difficult to implement new algorithms based on MOEA/D. With these frameworks, it is necessary to reimplement a complete algorithm (with code duplication) to implement variants of MOEA/D whereas with our framework, it is only sufficient to implement the component or the function that we want to modify or study. Each of the examples proposed in the documentation cannot therefore be implemented more easily than with our framework.

Here you are trying to explain the modularity that the package offers. I understand.

Although, in the paper pymoo, pygmo and jMetal are mentioned followed by the sentence "..._but they do not enable to test and analyze in detail the behavior of components implemented in each algorithm_"

Either you have to show somehow the aforementioned phrase or you should rephrase.
