### 2021-09-25 f3993c3
Revert "add 2 new full examples"
