### 2021-09-25 moead-framework/framework#71 @geoffreyp
New examples are available in the PR : #72 (merged by error when the issue https://github.com/moead-framework/framework/issues/69 was closed)
The user guide is available in the PR #73
The contributing page is updated in the PR #73

For the last check :

> The code coverage is great 95%. Although, as I can see the tests cover only the happy path. For instance if you pass not expected types the user should guess what is happening.

do you have a recommendation for this? Do I have to test each type of parameter, even if types are specified in the documentation?
