### 2021-09-25 moead-framework/framework#70 @geoffreyp
The paper is updated in https://github.com/moead-framework/framework/pull/76

> These implementations offer many state-of-the-art algorithms, visualization tools or parallelization abstraction, but they are not modular enough to test easily all MOEA/D components.
