### 2021-11-12 #2974 @geoffreyp
Hello @kyleniemeyer, I am waiting for the feedback from @chkoar (or from @sjvrijn if he has more time than @chkoar) ?
We still have 2 issues to close :
- https://github.com/moead-framework/framework/issues/71
- https://github.com/moead-framework/framework/issues/70
