### 2021-12-06 moead-framework/framework#71 @chkoar
> do you have a recommendation for this?

[hypothesis](https://hypothesis.readthedocs.io) (and [mutmut](https://mutmut.readthedocs.io)) might help you in this regard.

> Do I have to test each type of parameter, even if types are specified in the documentation?

Since you are in the process of publishing the library I think that you should validate each parameter (for value or/and type) and raise a human friendly error message.
