### 2021-12-06 moead-framework/framework#70 @chkoar
>  but they are not modular enough to test easily all MOEA/D components.

Should we show somehow this claim?
