### 2021-12-08 moead-framework/framework#70 @geoffreyp
I don't know, I don't see how to demonstrate otherwise that if the other libraries are not modular then we can't test specific components easily without completely reimplementing the algorithm ?
