### 2021-12-15 moead-framework/framework#70 @chkoar
> I don't know, I don't see how to demonstrate otherwise that if the other libraries are not modular then we can't test specific components easily without completely reimplementing the algorithm ?

Could you make a comparison table of moead-framwork with SOTA libraries. Like this you can show a direct comparison without implementing a complete example.
