### 2021-12-17 3cc456f
test types of MOEA/D and Problem parameters
