+++ moead_framework/problem/combinatorial/knapsack.py
@@ -25,13 +25,20 @@ class KnapsackProblem(Problem):
     >>> # Print all objectives values of the solution
     >>> print(solution.F)
     """
-    def __init__(self, number_of_objective, instance_file):
+    def __init__(self, number_of_objective: int, instance_file: str):
         """
         Constructor of the problem
 
         :param number_of_objective: {integer}
         :param instance_file: {string} txt file of the instance: http://www-desir.lip6.fr/~lustt/Research.html#MOKP
         """
+
+        if not isinstance(instance_file, str):
+            raise TypeError("The expected type of `instance_file` is `str`")
+
+        if not isinstance(number_of_objective, int):
+            raise TypeError("The expected type of `number_of_objective` is `int`")
+
         super().__init__(number_of_objective)
 
         self.instance_file = instance_file
@@ -67,7 +74,7 @@ class KnapsackProblem(Problem):
         self.number_of_objects = len(self.weights[0])
         file.close()
 
-    def f(self, function_id, decision_vector):
+    def f(self, function_id: int, decision_vector: np.ndarray):
         """
         Evaluate the decision_vector for the objective function_id
 
@@ -75,6 +82,12 @@ class KnapsackProblem(Problem):
         :param decision_vector: {:class:`~moead_framework.solution.one_dimension_solution.OneDimensionSolution`} solution to evaluate
         :return: {float} fitness value
         """
+
+        if not isinstance(function_id, int):
+            raise TypeError("The expected type of `function_id` is `int`")
+
+        if not isinstance(decision_vector, np.ndarray):
+            raise TypeError("The expected type of `decision_vector` is `np.ndarray`")
 
         function_id = function_id - 1
         weight = self.weight_of_solution(function_id, decision_vector)
