+++ moead_framework/problem/combinatorial/rmnk.py
@@ -30,12 +30,15 @@ class Rmnk(Problem):
 
     dtype = float
 
-    def __init__(self, instance_file):
+    def __init__(self, instance_file: str):
         """
         Constructor of the problem
 
         :param instance_file: {string} text file generated by the rmnk generator : http://mocobench.sourceforge.net/index.php?n=Problem.RMNK#Code
         """
+        if not isinstance(instance_file, str):
+            raise TypeError("The expected type of `instance_file` is `str`")
+
         file = open(instance_file, 'r')
         file_content = list(map(str.strip, file.readlines()))
         file_content = file_content[3:]
@@ -61,7 +64,7 @@ class Rmnk(Problem):
 
         file.close()
 
-    def f(self, function_id, decision_vector):
+    def f(self, function_id: int, decision_vector: np.ndarray):
         """
         Evaluate the decision_vector for the objective function_id
 
@@ -69,6 +72,11 @@ class Rmnk(Problem):
         :param decision_vector: {:class:`~moead_framework.solution.one_dimension_solution.OneDimensionSolution`} solution to evaluate
         :return: {float} fitness value
         """
+        if not isinstance(function_id, int):
+            raise TypeError("The expected type of `function_id` is `int`")
+
+        if not isinstance(decision_vector, np.ndarray):
+            raise TypeError("The expected type of `decision_vector` is `np.ndarray`")
 
         accu = 0
 
