### 2021-12-17 d814e88
a comparison table of moead-framework with SOTA libraries
