### 2022-01-12 moead-framework/framework#71 @chkoar
```
def test_that_zdt1_size():
    from moead_framework.problem.numerical import Zdt1
    Zdt1(1)
```
