### 2022-01-12 1c0256a
Refactor genetic operators
