+++ moead_framework/core/genetic_operator/numerical/moead_de_operators.py
@@ -1,9 +1,12 @@
 import random
 import numpy as np
 from moead_framework.core.genetic_operator.abstract_operator import GeneticOperator
-from moead_framework.core.genetic_operator.numerical.differential_evolution_crossover import \
-    DifferentialEvolutionCrossover
-from moead_framework.core.genetic_operator.numerical.polynomial_mutation import PolynomialMutation
+from moead_framework.core.genetic_operator.numerical.differential_evolution_crossover import (
+    DifferentialEvolutionCrossover,
+)
+from moead_framework.core.genetic_operator.numerical.polynomial_mutation import (
+    PolynomialMutation,
+)
 
 
 class MoeadDeOperators(GeneticOperator):
@@ -25,13 +28,10 @@ class MoeadDeOperators(GeneticOperator):
         solution2 = self.solutions[1]
         solution3 = self.solutions[2]
 
-        child_cross = DifferentialEvolutionCrossover(solutions=[solution1,
-                                                                solution2,
-                                                                solution3]).run()
+        child_cross = DifferentialEvolutionCrossover(
+            solutions=[solution1, solution2, solution3]
+        ).run()
 
         child_mut = PolynomialMutation(solutions=[child_cross]).run()
 
         return child_mut
-
-
-
