+++ moead_framework/core/genetic_operator/numerical/polynomial_mutation.py
@@ -19,7 +19,7 @@ class PolynomialMutation(GeneticOperator):
         self.number_of_solution_is_correct(n=1)
         solution = self.solutions[0]
 
-        return self.mutation(s=solution, rate=1/len(solution), n=len(solution))
+        return self.mutation(s=solution, rate=1 / len(solution), n=len(solution))
 
     def repair(self, s, mini=0, maxi=1):
         """
@@ -44,7 +44,12 @@ class PolynomialMutation(GeneticOperator):
         :return:
         """
         rand = random.uniform(0, 1)
-        return self.repair([s[x] if rand > rate else s[x] + self.sigma(n) * (maxi - (mini)) for x in range(len(s))])
+        return self.repair(
+            [
+                s[x] if rand > rate else s[x] + self.sigma(n) * (maxi - (mini))
+                for x in range(len(s))
+            ]
+        )
 
     def sigma(self, n):
         """
