### 2022-01-14 ec8bac7
Update moead_framework/problem/combinatorial/knapsack.py
Co-authored-by: Christos Aridas <chkoar@users.noreply.github.com>
