### 2022-01-14 0b3ea45
Update moead_framework/problem/combinatorial/knapsack.py
Co-authored-by: Christos Aridas <chkoar@users.noreply.github.com>
