### 2022-01-14 b9201d4
test parameters of init_with_data
