### 2022-03-05 moead-framework/framework#71 @geoffreyp
I added unit tests for the unhappy/sad path in the PR https://github.com/moead-framework/framework/pull/77.
All PRs (the last 4) are merged in the branch [review](https://github.com/moead-framework/framework/tree/review).
