### 2022-03-26 moead-framework/framework#71 @chkoar
Great thank you.

The following seem to raise exceptions but they are not tested. Could you please add them?
```
GeneticOperator.number_of_solution_is_correct
DeltaSelector.__init__
SpsDra.get_sub_problems
SpsRandomAndBoundaries.get_sub_problems
```

In general I think that all components that are listed for public consumption should be properly tested for behavior and exceptions. So, the user should be informed accordingly.  In general do not trust users input. Some examples, for instance.
```python
# test/test_genetic_operator.py

# I am allowed to pass a string in the mutation but when I run the mutation the message is not informative
# Result: TypeError: unsupported operand type(s) for -: 'str' and 'int'
def test_binary_mutattion_example1():
    array_solution = ["a",0]
    new_solution = BinaryMutation(solutions=[array_solution]).run()

# Should this be allowed?
# Result: [0.10000000000000009, 0.3999999999999999]
def test_binary_mutattion_example2():
    array_solution = [1.1, 1.4]
    new_solution = BinaryMutation(solutions=[array_solution]).run()

# Tried to pass a single array
# Result: TypeError: object of type 'int' has no len()
def test_binary_mutattion_example3():
    array_solution = [0, 1]
    new_solution = BinaryMutation(solutions=array_solution).run()

# Tried to pass an integer
# Result: TypeError: 'int' object is not subscriptable
def test_binary_mutattion_example4():
    array_solution = [0, 1]
    new_solution = BinaryMutation(solutions=array_solution).run()

# Tried to pass empty arrays
# Result: ZeroDivisionError: division by zero
def test_binary_mutattion_example5():
    array_solution = [0, 1]
    new_solution = BinaryMutation(solutions=[[]]).run()

# Tried to ask more crossover points than the length
# Result: Infinite loop
def test_crossover_example1():
    array_solution1 = [1, 2, 3, 4, 5]
    array_solution2 = [6, 7, 8, 9, 10]
    new_solution_one_point = Crossover(solutions=[array_solution1, array_solution2],crossover_points=10).run().tolist()

# Tried to ask a zero crossover points than the length
# Result: IndexError: list index out of range
def test_crossover_example1():
    array_solution1 = [1, 2, 3, 4, 5]
    array_solution2 = [6, 7, 8, 9, 10]
    new_solution_one_point = Crossover(solutions=[array_solution1, array_solution2],crossover_points=0).run().tolist()

```
