### 2022-04-13 f9ff741
add more unit test : sad path
