+++ moead_framework/test/test_genetic_operator.py
@@ -73,3 +73,49 @@ class GeneticOperatorsTest(unittest.TestCase):
 
         self.assertEqual(new_solution, [0.021382317860195066, 1, 0.7327909561319956,
                                         0.17080281343876919, 0.2590886275388289])
+
+    def test_binary_mutation_example1(self):
+        array_solution = ["a", 0]
+        with self.assertRaises(AttributeError):
+            new_solution = BinaryMutation(solutions=[array_solution]).run()
+
+    # Tried to pass a single array
+    # Result: TypeError: object of type 'int' has no len()
+    def test_binary_mutation_example3(self):
+        array_solution = [0, 1]
+        with self.assertRaises(AttributeError):
+            new_solution = BinaryMutation(solutions=array_solution).run()
+
+    # Tried to pass an integer
+    # Result: TypeError: 'int' object is not subscriptable
+    def test_binary_mutation_example4(self):
+        array_solution = [0, 1]
+        with self.assertRaises(AttributeError):
+            new_solution = BinaryMutation(solutions=array_solution).run()
+
+    # Tried to pass empty arrays
+    # Result: ZeroDivisionError: division by zero
+    def test_binary_mutation_example5(self):
+        array_solution = [0, 1]
+        with self.assertRaises(AttributeError):
+            new_solution = BinaryMutation(solutions=[[]]).run()
+
+    # Tried to ask more crossover points than the length
+    # Result: Infinite loop
+    def test_crossover_example(self):
+        array_solution1 = [1, 2, 3, 4, 5]
+        array_solution2 = [6, 7, 8, 9, 10]
+
+        with self.assertRaises(ValueError):
+            new_solution_one_point = Crossover(solutions=[array_solution1, array_solution2],
+                                               crossover_points=10).run().tolist()
+
+    # Tried to ask a zero crossover points than the length
+    # Result: IndexError: list index out of range
+    def test_crossover_example1(self):
+        array_solution1 = [1, 2, 3, 4, 5]
+        array_solution2 = [6, 7, 8, 9, 10]
+
+        with self.assertRaises(ValueError):
+            new_solution_one_point = Crossover(solutions=[array_solution1, array_solution2],
+                                               crossover_points=0).run().tolist()
