### 2022-08-05 #2974 @arfon
Hi folks, just checking in again here. @chkoar @geoffreyp – it looks like you're finessing tests in https://github.com/moead-framework/framework/issues/71 ? Having reviewed that thread I would recommend moving forward with an accept/publish at this point @melissawm.
