### 2022-08-22 #2974 @chkoar
> @sjvrijn @chkoar are you happy with us moving to accept and publish this paper at this point?

Yeap
