+++ docs/moead_framework/moead_framework.problem.combinatorial.rmnk.Rmnk.rst
@@ -17,6 +17,8 @@ moead\_framework.problem.combinatorial.rmnk.Rmnk
       ~Rmnk.evaluate
       ~Rmnk.f
       ~Rmnk.generate_random_solution
+      ~Rmnk.init_with_data
+      ~Rmnk.init_with_instance_file
       ~Rmnk.load_links
       ~Rmnk.load_tables
       ~Rmnk.sigma
