+++ docs/moead_framework/moead_framework.problem.numerical.zdt.Zdt1.rst
@@ -17,7 +17,6 @@ moead\_framework.problem.numerical.zdt.Zdt1
       ~Zdt1.evaluate
       ~Zdt1.f
       ~Zdt1.generate_random_solution
-      ~Zdt1.generate_solution
       ~Zdt1.validate_size
    
    
