### 2022-10-16 395a658
Minor typo fixes
