### 2022-10-16 #2974 @arfon
@geoffreyp – I found a couple of minor typos in your paper here: https://github.com/moead-framework/framework/pull/79
